% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataReformatting.R
\name{add_column_after_specific_column}
\alias{add_column_after_specific_column}
\title{Add column after specific column}
\usage{
add_column_after_specific_column(
  data,
  new_data,
  new_column_name = NULL,
  old_column_name,
  loop_structure
)
}
\arguments{
\item{data}{The data containing the original columns}

\item{new_data}{The new data which we want to include}

\item{new_column_name}{The column name (or name pattern for loops) for the
new data we would like to add}

\item{old_column_name}{The marker column name (or name pattern for loops)
where we want to add the new data}

\item{loop_structure}{A boolean indicating whether this information is
in a loop format (TRUE) or a single pair of variables (FALSE)}
}
\value{

}
\description{
During the data processing, some new columns need
to be added, with unit conversions or newly calculated
values. This function allows us to add those columns
in specific locations to make analysis easier.
}
