% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MongoDB.R
\name{add_data_to_db}
\alias{add_data_to_db}
\title{Add data to db
  
Rpackage file: MongoDB.R}
\usage{
add_data_to_db(
  data,
  collection = "data",
  data_type,
  database = "rhomis",
  url = "mongodb://localhost",
  projectID,
  formID,
  overwrite = F
)
}
\arguments{
\item{data}{The tibble dataset that you would like to write}

\item{collection}{The collection which you would like to add the data to}

\item{data_type}{The type of data which is being added (e.g. indicatorData, processedData,metaData...)}

\item{database}{The database where the data is going to be saved}

\item{url}{The url of the database you are saving the database to. The default mongoDB url is used as the default here}

\item{projectID}{The ID of the project you would like to save}

\item{formID}{The ID of the form you would like to save}

\item{overwrite}{Whether or not to overwrite a previously saved version of the project}
}
\description{
Add data to db
  
Rpackage file: MongoDB.R
}
