% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetaData.R
\name{add_function_to_list}
\alias{add_function_to_list}
\title{Add function to List}
\usage{
add_function_to_list(function_list, function_name, called_by)
}
\arguments{
\item{function_list}{The list of functions}

\item{function_name}{The name of the function}

\item{called_by}{What this function is called by}
}
\description{
Add function to List
}
