% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetaData.R
\name{add_gendered_indicator}
\alias{add_gendered_indicator}
\title{Add Gendered indicator}
\usage{
add_gendered_indicator(
  indicator_list,
  gendered_list = c("male_youth", "male_adult", "female_youth", "female_adult"),
  indicator_name = "male_youth_crop_consumed_kg_per_year",
  output_format,
  individual_columns_required = list(),
  loop_columns_required = list(),
  conversion_tables_required = list(),
  api_data_required = list(),
  indicators_required = list(),
  function_calculated = "crop_gender_calculations"
)
}
\arguments{
\item{indicator_list}{The Existing list of indicators}

\item{gendered_list}{List of gender categories}

\item{indicator_name}{The name of the new indicator (without gender prefix)}

\item{output_format}{The shape of the indicator (column, loop or table)}

\item{individual_columns_required}{Which individual columns are required (directly) for the RHoMIS dataset}

\item{loop_columns_required}{Which looped columns are required}

\item{conversion_tables_required}{Which conversion tables are required}

\item{api_data_required}{Which api connection data is required}

\item{indicators_required}{Other indicators which are required to calculate this one}

\item{function_calculated}{The function where this indicator is calculated}
}
\description{
Add Gendered indicator
}
