% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenerateMockResponses.R
\name{add_headers_and_footers}
\alias{add_headers_and_footers}
\title{Add Headers and Footers}
\usage{
add_headers_and_footers(metadata, xml_string, survey_path = NULL)
}
\arguments{
\item{metadata}{The "settings" sheet of the xls form}

\item{xml_string}{The string which needs headers and
footers added}

\item{survey_path}{The path to the survey file where we
obtain the metadata information}
}
\value{

}
\description{
Submissions to ODK central have headers
and footers so the forms can be managed and metadata included.
This function adds the necessary headers and footer based on the survey file
}
