% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetaData.R
\name{add_indicator}
\alias{add_indicator}
\title{Add Indicator}
\usage{
add_indicator(
  indicator_list,
  indicator_name,
  output_format = c("column", "loop", "table"),
  individual_columns_required = list(),
  loop_columns_required = list(),
  conversion_tables_required = list(),
  api_data_required = list(),
  indicators_required = list(),
  function_calculated,
  search_term
)
}
\description{
Add an indicator for 
the RHoMIS dataset
}
