% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenerateMockResponses.R
\name{adding_looped_data}
\alias{adding_looped_data}
\title{Adding Looped Data}
\usage{
adding_looped_data(survey, choices, rows)
}
\arguments{
\item{survey}{The survey table as a tibble}

\item{choices}{The choices table as a tibble}

\item{rows}{The rows in the survey table which correspond
to the looped questions}
}
\description{
Some of the survey questions in ODK central are asked in a
repeat format. This function allows us to carry out these
repeats for a random number of loops (capped at 5)
}
\details{
Rpackage file: GenerateMockResponses.R
}
