% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MongoDB.R
\name{adding_project_to_list}
\alias{adding_project_to_list}
\title{Add Project to List}
\usage{
adding_project_to_list(
  database = "rhomis",
  url = "mongodb://localhost",
  projectID,
  formID
)
}
\arguments{
\item{database}{Database to add it to}

\item{url}{URL of the database}

\item{projectID}{ID of the project you are adding}

\item{formID}{The id of the form being you are adding}
}
\description{
After conducting the main RHoMIS calculations, it is important to add the
form ID and project ID to the list of projects in the database, to keep track of the
projects using RHoMIS. This function allows us to do this.
  
Rpackage file: MongoDB.R
}
