% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataReformatting.R
\name{apply_conversion_factor_to_columns}
\alias{apply_conversion_factor_to_columns}
\title{Apply Conversion Factor to Columns}
\usage{
apply_conversion_factor_to_columns(data, conversion_tibble, id_rhomis_dataset)
}
\arguments{
\item{data}{A tibble to be converted}

\item{conversion_tibble}{The conversion tibble}

\item{id_rhomis_dataset}{The id of the project to
be converted}
}
\description{
Sometimes it can be helpful to apply conversion
factors to a series of columns, for example to
get the calorie values of crops produced. This
function allows us to apply conversion factors
to an individual project.
}
\details{
Rpackage file: DataReformatting.R
}
