% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataReformatting.R
\name{apply_conversion_factor_to_columns_multiple_projects}
\alias{apply_conversion_factor_to_columns_multiple_projects}
\title{Apply Conversion Factor to Columns Multiple Projects}
\usage{
apply_conversion_factor_to_columns_multiple_projects(
  data,
  conversion_tibble,
  by_project = T
)
}
\arguments{
\item{data}{A tibble to be converted}

\item{conversion_tibble}{The conversion tibble}

\item{by_project}{Whether or not to apply
conversion factors by project
be converted}
}
\description{
Sometimes it can be helpful to apply conversion
factors to a series of columns, for example to
get the calorie values of crops produced. This
function allows us to apply conversion factors
to mulptiple projects.
}
\details{
Rpackage file: DataReformatting.R
}
