% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LivestockCalculations.R
\name{average_good_and_bad_season}
\alias{average_good_and_bad_season}
\title{Calculate Average Milk Harvested Values}
\usage{
average_good_and_bad_season(good_season_amount, bad_season_amount)
}
\arguments{
\item{good_season_amount}{Vector of good season yields}

\item{bad_season_amount}{Vector of bad season yields}
}
\description{
Averaging yield for the good season and bad season
Where one season is NA, we only use information for the season available.
Where both seasons are NA, the final is calculated as NA
  
Rpackage file: LivestockCalculations.R
}
