% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NTFP.R
\name{calculate_fp_harvest}
\alias{calculate_fp_harvest}
\title{Calculate fp Harvest}
\usage{
calculate_fp_harvest(
  tree_aid_df,
  fp_harvest_conversions,
  name_column,
  amount_column,
  unit_column
)
}
\arguments{
\item{tree_aid_df}{A tree aid data frame}

\item{fp_harvest_conversions}{The conversion table for NTFP harvests}

\item{name_column}{The column containing the name of the product}

\item{amount_column}{The column containing the amount}

\item{unit_column}{The column containing the survey unit}
}
\description{
Calculate the amount of forest products harvested in KG
}
