% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HouseholdSize.R
\name{calculate_household_size_members}
\alias{calculate_household_size_members}
\title{Calculate Household Size Members}
\usage{
calculate_household_size_members(data)
}
\arguments{
\item{data}{RHoMIS data with household roster}
}
\value{

}
\description{
Calculate Household Size based on roster
questions
}
