% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-calculate-indicators.R
\name{calculate_indicators_local}
\alias{calculate_indicators_local}
\title{Calculate all of the Local Indicators}
\usage{
calculate_indicators_local(
  base_path = "./",
  file_path,
  id_type = c("string", "column"),
  proj_id,
  form_id,
  gender_categories = pkg.env$gender_categories,
  unique_id_col = "_uuid"
)
}
\arguments{
\item{unique_id_col}{The column containing unique household ids}
}
\description{
Calculate all of the Local Indicators
}
