% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-calculate-prices.R
\name{calculate_prices_csv}
\alias{calculate_prices_csv}
\title{Calculate Prices CSV}
\usage{
calculate_prices_csv(project_folder = "./")
}
\arguments{
\item{project_folder}{}
}
\description{
Rpackage file: 02-calculate-prices.R
ß
}
