% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redirectModules.R
\name{calculate_values_gender_and_fa_local}
\alias{calculate_values_gender_and_fa_local}
\title{Calculate Values, Gender and Food Availability}
\usage{
calculate_values_gender_and_fa_local(
  base_path = "./",
  processed_data_path = "processed_data/",
  indicator_path = "indicator_data/",
  units_path = "converted_units/",
  calories_path = "completed_calorie_conversions/",
  prices_path = "converted_prices/",
  staple_crop = "maize",
  gender_categories = c("male_youth", "female_youth", "male_adult", "female_adult")
)
}
\arguments{
\item{base_path}{Path to the project folder}

\item{processed_data_path}{Path to processed data folder}

\item{indicator_path}{Path to indicator folder}

\item{units_path}{Path to the folder of converted units}

\item{calories_path}{Path to calories folder}

\item{prices_path}{Path to prices folder}

\item{staple_crop}{The main staple crop consumed in the area}
}
\value{

}
\description{
Calculate Values, Gender and Food Availability
}
