% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataContentsCheck.R
\name{check_columns_in_data}
\alias{check_columns_in_data}
\title{Check columns in data}
\usage{
check_columns_in_data(data, loop_columns = NULL, individual_columns = NULL)
}
\arguments{
\item{data}{The dataset which needs checking}

\item{loop_columns}{The columns which are contained in a loop}

\item{individual_columns}{The columns which appear individually}
}
\value{

}
\description{
For many of the functions to work, we need to ensure that
the correct columns are present in the tibble. This
function checks whether the columns are present, both
for looped, and individual columns
}
