% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataContentsCheck.R
\name{check_columns_individual}
\alias{check_columns_individual}
\title{Check Columns Individual}
\usage{
check_columns_individual(data, columns)
}
\arguments{
\item{data}{The data we are checking}

\item{columns}{The column names we are looking for}
}
\description{
Checking whether individual columns are present in the data
we are examining
}
\details{
Rpackage file: DataContentsCheck.R
}
