% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataContentsCheck.R
\name{check_columns_loop}
\alias{check_columns_loop}
\title{Check Columns Loop}
\usage{
check_columns_loop(data, columns)
}
\arguments{
\item{data}{Data to be checked}

\item{columns}{Column patterns that we are verifying}
}
\value{

}
\description{
Check whether columns are present in loop type data
}
