% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UnitsAndConversions.R
\name{check_existing_calorie_conversions}
\alias{check_existing_calorie_conversions}
\title{Check Existing Crop Calorie Conversion}
\usage{
check_existing_calorie_conversions(list_of_df)
}
\arguments{
\item{list_of_df}{A list of dataframes, containing all
of the units and conversions}
}
\value{

}
\description{
Go through the common conversions stored in the R-package
}
