% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReformatLoops.R
\name{check_val_in_list}
\alias{check_val_in_list}
\title{Check Value is in List}
\usage{
check_val_in_list(item, category)
}
\arguments{
\item{item}{A list of gender control strings (see example)}

\item{category}{The category you are trying to identify}
}
\value{

}
\description{
In the case of the gender calculations, this function can check whether
a particular gender category is found in a string.
}
\examples{

item <- c("male_adult female_adult", "male_adult", NA, "female_youth")
category <- "female_youth"
check_val_in_list(item, category)

item <- c("male_adult female_adult", "male_adult", NA, "female_youth")
category <- "male_adult"
check_val_in_list(item, category)
}
