% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CleanColumnNames.R
\name{clean_column_names}
\alias{clean_column_names}
\title{Clean Column Names}
\usage{
clean_column_names(column_names, seperator, repeat_columns)
}
\arguments{
\item{column_names}{A list of all of the column names}

\item{seperator}{The seperator which seperates all of the parts of the column names}

\item{repeat_columns}{A list of all of the repeat loops contained in the RHoMIS dataset}
}
\value{
A cleaned list of all of the column names
}
\description{
A single functions to shorten and clean all RHoMIS column names.
}
\examples{

}
