% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MongoDB.R
\name{clean_json_string}
\alias{clean_json_string}
\title{Clean JSON String}
\usage{
clean_json_string(json_string)
}
\arguments{
\item{json_string}{The json string which needs to be cleaned}
}
\value{

}
\description{
The strings produced from "json::stringify", when
processed, can contain some unnecessary characters.
}
\details{
This function can be useful for removing these unnecessary
characters.
}
