% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataReformatting.R
\name{collapse_list_of_tibbles}
\alias{collapse_list_of_tibbles}
\title{Collapse list of tibbles}
\usage{
collapse_list_of_tibbles(list_of_tibbles)
}
\arguments{
\item{list_of_tibbles}{A list of tibbles which need to be collapsed}
}
\value{

}
\description{
A useful function for collapsing a list of tibbles,
used predominantly for gender calculations
}
