% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MongoDB.R
\name{connect_to_db}
\alias{connect_to_db}
\title{Connect to db}
\usage{
connect_to_db(collection, database = "rhomis", url = "mongodb://localhost")
}
\arguments{
\item{collection}{Which collection you want to connect to}

\item{database}{Which database you want to connect to}

\item{url}{The URL of the database you want to connect to}
}
\value{
Returns a connection object
}
\description{
A simple function for connecting to a mongo database
}
