% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CleanRawData.R
\name{convert_all_columns_to_lower_case}
\alias{convert_all_columns_to_lower_case}
\title{Convert All Columns to Lower Case}
\usage{
convert_all_columns_to_lower_case(data)
}
\arguments{
\item{data}{A tibble to be converted to lower case}
}
\description{
Convert all character columns in the data set
to lower case
}
\details{
Rpackage file: CleanRawData.R
}
