% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CurrencyConversion.R
\name{convert_all_currencies}
\alias{convert_all_currencies}
\title{Convert all Currencies}
\usage{
convert_all_currencies(data, country_column = "country", year_column = "year")
}
\arguments{
\item{data}{The data containing a column of ISO country codes and
a column of Years}

\item{country_column}{The name of the country column}

\item{year_column}{The name of the year column}
}
\value{

}
\description{
A function for finding the relevant currency conversion
factors for all household interviews
}
