% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CleanRawData.R
\name{convert_column_to_lowercase}
\alias{convert_column_to_lowercase}
\title{Convert Column to Lowercase}
\usage{
convert_column_to_lowercase(column)
}
\arguments{
\item{column}{Column to be converted.}
}
\value{

}
\description{
Convert an individual column to lower case
if it is a character column
}
