% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CropCalculations.R
\name{convert_crop_sold_units}
\alias{convert_crop_sold_units}
\title{Convert Crop Sold Units}
\usage{
convert_crop_sold_units(
  data,
  units = crop_price_units$unit,
  unit_conversions = crop_price_units$conversion
)
}
\arguments{
\item{data}{RHoMIS data containing crop-looping information}

\item{units}{A list of units to be converted}

\item{unit_conversions}{The numeric conversion factors for the units listed above}
}
\value{

}
\description{
Convert crop income units into a numeric conversion factor
}
