% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CropCalculations.R
\name{convert_crop_yield_units}
\alias{convert_crop_yield_units}
\title{Convert Crop Yield Units}
\usage{
convert_crop_yield_units(data, unit_conv_tibble = NULL)
}
\arguments{
\item{data}{The data containing the units to convert}

\item{units_conversion_tibble}{Unit tibble for crop yield, including individual ids for projects. Crops to be coverted to kg amount}
}
\value{

}
\description{
RHoMIS crop yields come in a character format. This function
converts these to numeric and adds the new numeric units in an
appropriate place for the dataset
}
