% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExtractNewValues.R
\name{convert_new_values_to_tibble}
\alias{convert_new_values_to_tibble}
\title{Convert New Values to Tibble}
\usage{
convert_new_values_to_tibble(new_values)
}
\arguments{
\item{new_values}{A list of lists containing the values to convert}
}
\description{
The output of the "extract_new_core_units" function
is a list of lists. With the names of the lists
are the variables to be converted and the lists
contain all of the new values.
This function converts all of these lists into tibbles.
With a new column which has to be filled
in by the user
}
\details{
Rpackage file: ExtractNewValues.R
}
\examples{

new_values <- c("egg1", "egg2", "egg3")
expected_result <- tibble::as_tibble(list(
  "survey_value" = c("egg1", "egg2", "egg3"),
  "conversion" = c(NA, NA, NA)
))

actual_result <- convert_new_values_to_tibble(new_values)
}
