% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NTFP.R
\name{convert_ntfp_units}
\alias{convert_ntfp_units}
\title{Convert NTFP Units}
\usage{
convert_ntfp_units(unit_data, units_conversions)
}
\arguments{
\item{unit_data}{The columns in the raw data containing units}

\item{units_conversions}{The list of unit conversion tibbles}
}
\description{
Switch Out NTFP units for numeric
conversion factors
}
