% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MongoDB.R
\name{count_collection}
\alias{count_collection}
\title{Count collection}
\usage{
count_collection(collection, database = "rhomis", url = "mongodb://localhost")
}
\arguments{
\item{collection}{The collection you want to collect the information from}

\item{database}{The database you are querying}

\item{url}{The url of the database you are querying}
}
\value{

}
\description{
Count the number of documents in a collection
}
