% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{country}
\alias{country}
\title{Country}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 34 rows and 2 columns.
}
\usage{
country
}
\description{
A dataset which converts country names found in previous RHoMIS surveys
into two letter ISO country codes
}
\details{
@source NA
}
\keyword{datasets}
