% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{country_codes}
\alias{country_codes}
\title{ISO Country Codes}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 249 rows and 11 columns.
}
\source{
iso
}
\usage{
country_codes
}
\description{
Conversions for official names
country/region codes
}
\keyword{datasets}
