% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{country_to_iso2}
\alias{country_to_iso2}
\title{Country Conversions to ISO}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 40 rows and 2 columns.
}
\usage{
country_to_iso2
}
\description{
A dataset which converts country names found in previous RHoMIS surveys
into two letter ISO country codes
}
\details{
@source NA
}
\keyword{datasets}
