% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Modules.R
\name{create_module}
\alias{create_module}
\title{Create a new RHoMIS module with Analysis scripts}
\usage{
create_module(module_name, module_type, processing_code, dependencies)
}
\arguments{
\item{module_name}{The name of the module you would like to add}

\item{module_type}{The type of module to add (e.g. core,optional_module, tailored)}

\item{processing_code}{The code used to process the modules}

\item{dependencies}{Which modules this module depends on}
}
\description{
Create a module to save in the RHoMIS database
  
Rpackage file: Modules.R
}
