% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{crop_amount_to_kg}
\alias{crop_amount_to_kg}
\title{Crop Yield Units}
\format{
A tibble
}
\source{
NA
}
\usage{
crop_amount_to_kg
}
\description{
Conversion factors for the most common
crop yield units in the RHoMIS dataset
}
\keyword{datasets}
