% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{crop_calories}
\alias{crop_calories}
\title{Crop Calories}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 111 rows and 2 columns.
}
\source{
NA
}
\usage{
crop_calories
}
\description{
A list of calorie conversions for common crops
}
\keyword{datasets}
