% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CropAndLivestockDiversity.R
\name{crop_diversity}
\alias{crop_diversity}
\title{Crop Diversity}
\usage{
crop_diversity(data, indicator_data)
}
\arguments{
\item{data}{RHoMIS dataset}

\item{indicator_data}{A tibble to store indicators, if crop diversity can be calculated, then the indicator will be appended.}
}
\value{

}
\description{
Calculating the number of crops kept by a household
}
