% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CropCalculations.R
\name{crop_gender_calculations}
\alias{crop_gender_calculations}
\title{Crop Gender Calculations}
\usage{
crop_gender_calculations(data)
}
\arguments{
\item{data}{RHoMIS data set conatining processed values for
crop consumed, crop sold, and crop income}
}
\value{

}
\description{
Adding the gendered information for crop consumed
crop sold, and crop income. These have to already have been
calculated in the dataset being used
}
