% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CropYield.R
\name{crop_harvest_calculations}
\alias{crop_harvest_calculations}
\title{Crop harvest calculations}
\usage{
crop_harvest_calculations(data)
}
\arguments{
\item{data}{RHoMIS data containing cropping information}
}
\value{

}
\description{
Calculating the amount of each crop harvested per year
}
