% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CropYield.R
\name{crop_harvest_single_loop}
\alias{crop_harvest_single_loop}
\title{Crop yield single loop}
\usage{
crop_harvest_single_loop(data, loop_number)
}
\arguments{
\item{data}{THe data containind crop yields, and numeric
crop yield units}

\item{loop_number}{which loop you are calculating crop yield for.}
}
\value{

}
\description{
Calculate crop yield based off of a single loop
}
