% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{crop_name}
\alias{crop_name}
\title{Crop Name}
\format{
An object of class \code{character} of length 21.
}
\usage{
crop_name
}
\description{
A list of common crop names which can be found in
previous rhomis datasets
}
\details{
@source NA
}
\keyword{datasets}
