% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{crop_name_to_std}
\alias{crop_name_to_std}
\title{Crop Name}
\format{
An object of class \code{character} of length 20.
}
\usage{
crop_name_to_std
}
\description{
A list of common crop names which can be found in
previous rhomis datasets
}
\details{
@source NA
}
\keyword{datasets}
