% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{crop_price_to_lcu_per_kg}
\alias{crop_price_to_lcu_per_kg}
\title{Crop Price Units}
\format{
A tibble
}
\source{
NA
}
\usage{
crop_price_to_lcu_per_kg
}
\description{
Conversion factors for the most common
crop price units in the RHoMIS dataset
}
\keyword{datasets}
