% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CropCalculations.R
\name{crop_proportions_all}
\alias{crop_proportions_all}
\title{Crop Proportions for all}
\usage{
crop_proportions_all(data)
}
\arguments{
\item{data}{A standard RHoMIS data set}
}
\description{
A function for calculating the numeric proportions of crops which are sold or
consumed
}
\details{
Rpackage file: CropCalculations.R
}
