% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CropYield.R
\name{crop_sold_and_consumed_calculation}
\alias{crop_sold_and_consumed_calculation}
\title{Crop sold and consumed.}
\usage{
crop_sold_and_consumed_calculation(data)
}
\arguments{
\item{data}{RHoMIS data which has been processed to include
crops harvested (see above) and crop proportions sold/consumed.}
}
\value{

}
\description{
Note that for this function to work,
previous functions have to have been run. The crops harvested has to have been
calculated, as well as the proportions (numeric) which have been sold
and consumed
}
