% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{currency_conversion}
\alias{currency_conversion}
\title{Currency Conversions}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 16492 rows and 6 columns.
}
\source{
World Bank
}
\usage{
currency_conversion
}
\description{
World Bank currency conversions
}
\keyword{datasets}
