% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CurrencyConversion.R
\name{currency_conversion_factor}
\alias{currency_conversion_factor}
\title{Conversion_factor}
\usage{
currency_conversion_factor(year, country_code)
}
\arguments{
\item{year}{The year you are querying the conversion factors for}

\item{country_code}{The country codes that you are querying for}
}
\value{

}
\description{
Identify the closest available currency conversion factor
based on world bank data. If we cannot find the current
currency conversion factor, we will check for the most recent,
going as far back as 10 years
}
