% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MongoDB.R
\name{data_frame_to_json}
\alias{data_frame_to_json}
\title{Data frame to JSON}
\usage{
data_frame_to_json(data_to_change)
}
\arguments{
\item{data_to_change}{The data frame that needs to be converted}
}
\value{
A JSON string
}
\description{
To write objects to a mongoDB, they need to be in JSON format.
This function converts data from a data frame to a JSON
}
\examples{
sample_data_frame <- as_tibble(list("original_spelling"=c("benana","maz","wetermalon","cokonut"),
"standardised_spelling"=c("benana",NA,"wetermalon","cokonut")))
data_frame_to_json(sample_data_frame)

}
