% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConnectToCentral.R
\name{delete_project}
\alias{delete_project}
\title{Delete Project}
\usage{
delete_project(central_url, central_email, central_password, projectID)
}
\arguments{
\item{central_url}{The url of the ODK central server}

\item{central_email}{The email of your ODK central account}

\item{central_password}{The password to your ODK central account}

\item{projectID}{The ID of the project you are hoping to delete}
}
\description{
Delete a project from ODK central based on its ID
}
\details{
Rpackage file: ConnectToCentral.R
}
