% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AuthenticateUser.R
\name{delete_user}
\alias{delete_user}
\title{Delete User}
\usage{
delete_user(url, email, password)
}
\arguments{
\item{url}{The URL of the authentication server}

\item{email}{The email of the account being deleted}

\item{password}{The password of the account being deleted}
}
\value{

}
\description{
A function to delete a user from the RHoMIS authentication server
}
