% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QualityChecking.R
\name{dependency_check}
\alias{dependency_check}
\title{Dependency Check}
\usage{
dependency_check(row_indexes, indicator, survey_data, indicator_data = NULL)
}
\arguments{
\item{row_indexes}{Rows you are interested in}

\item{indicator}{Indicator interested in}

\item{survey_data}{Survey Dataset}

\item{indicator_data}{Indicator sheet}
}
\description{
Check the the data for an indicator dependency
for a particular set of rows
}
