% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LivestockCalculations.R
\name{egg_income_calculations}
\alias{egg_income_calculations}
\title{Eggs Income Calculations}
\usage{
egg_income_calculations(data, unit_conv_tibble = NULL)
}
\arguments{
\item{data}{RHoMIS data with livestock loops included.}

\item{unit_conv_tibble}{A tibble containing common su}
}
\value{

}
\description{
Function to calculate egg income from livestock loops
}
