% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{eggs_amount_units}
\alias{eggs_amount_units}
\title{Eggs Amount Units}
\format{
A tibble
}
\source{
NA
}
\usage{
eggs_amount_units
}
\description{
Conversion factors for the amount
of eggs collected over a particular unit of time
}
\keyword{datasets}
