% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{eggs_calories}
\alias{eggs_calories}
\title{Egg Calories}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 12 rows and 2 columns.
}
\source{
NA
}
\usage{
eggs_calories
}
\description{
A list of calorie conversion for eggs
}
\keyword{datasets}
