% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LivestockCalculations.R
\name{eggs_price_per_egg_to_numeric}
\alias{eggs_price_per_egg_to_numeric}
\title{Eggs Price per Egg to Numeric}
\usage{
eggs_price_per_egg_to_numeric(units_column, amount_sold_column)
}
\arguments{
\item{units_column}{A vector of crop units to convert}

\item{amount_sold_column}{A vector of amounts of eggs sold (in kg)}
}
\description{
A function to make sure that price per egg can be converted
for the correct calculations to be made
  
Rpackage file: LivestockCalculations.R
}
