% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{eggs_price_time_units}
\alias{eggs_price_time_units}
\title{Eggs Price Time Units}
\format{
A tibble
}
\source{
NA
}
\usage{
eggs_price_time_units
}
\description{
Conversion factors for income from egg production
}
\keyword{datasets}
