% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{eggs_price_to_lcu_per_year}
\alias{eggs_price_to_lcu_per_year}
\title{Eggs Price Time Units}
\format{
A tibble
}
\source{
NA
}
\usage{
eggs_price_to_lcu_per_year
}
\description{
Conversion factors for income from egg production
}
\keyword{datasets}
