% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LivestockCalculations.R
\name{eggs_proportions_all}
\alias{eggs_proportions_all}
\title{Eggs Proportions All}
\usage{
eggs_proportions_all(data)
}
\arguments{
\item{data}{The RHoMIS data including livestock loops and
livestock heads data}
}
\value{

}
\description{
A function for calculating the proportions of
eggs and consumed
}
