% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LivestockCalculations.R
\name{eggs_sold_and_consumed_calculations}
\alias{eggs_sold_and_consumed_calculations}
\title{Eggs Sold and Consumed Calculations}
\usage{
eggs_sold_and_consumed_calculations(data)
}
\arguments{
\item{data}{Data containing livestock loops to calculate the amounts of
egg sold and consumed}
}
\value{

}
\description{
Function for calculating the amounts of eggs sold and consumed
}
