% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LivestockCalculations.R
\name{eggs_swap_per_animal_units}
\alias{eggs_swap_per_animal_units}
\title{Eggs Swap per Animal Units
  
Rpackage file: LivestockCalculations.R}
\usage{
eggs_swap_per_animal_units(
  units_column,
  livestock_name_column,
  livestock_heads_df
)
}
\arguments{
\item{units_column}{A vector containing the units to be converted}

\item{livestock_name_column}{A vector of livestock names}

\item{livestock_heads_df}{A tibble of number of livestock heads
with column names "livestock_heads_cattle","livestock_heads_chicken" etc...}
}
\description{
Eggs Swap per Animal Units
  
Rpackage file: LivestockCalculations.R
}
