% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunAll.R
\name{extract_all_new_values}
\alias{extract_all_new_values}
\title{Extract New Values}
\usage{
extract_all_new_values(data)
}
\arguments{
\item{data}{A RHoMIS tibble}
}
\description{
Extract all of the new values from a RHoMIS data frame,
if they have unit conversions in the package, then convert them.
 
Rpackage file: RunAll.R
}
