% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExtractNewValues.R
\name{extract_calorie_conversion_data_frames}
\alias{extract_calorie_conversion_data_frames}
\title{Extract Units to DataFrame}
\usage{
extract_calorie_conversion_data_frames(data)
}
\arguments{
\item{data}{The core RHoMIS dataset which we are extracting units from}
}
\value{
A list of data frames
}
\description{
Extract Units to DataFrame
}
