% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExtractNewValues.R
\name{extract_calorie_values}
\alias{extract_calorie_values}
\title{Extract names of Values to Receive Calorie Conversions}
\usage{
extract_calorie_values(data)
}
\arguments{
\item{data}{The RHoMIS data set conta}
}
\value{

}
\description{
Extract names of Values to Receive Calorie Conversions
}
