% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExtractNewValues.R
\name{extract_calorie_values_by_project}
\alias{extract_calorie_values_by_project}
\title{Extract Calorie Values by Project}
\usage{
extract_calorie_values_by_project(data)
}
\arguments{
\item{data}{RHoMIS Dataset}
}
\value{

}
\description{
A function to identify all of the crops and livestock
for which calorie conversion values need to be included for
}
