% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConnectToCentral.R
\name{extract_form_metadata}
\alias{extract_form_metadata}
\title{Extract Form Metadata}
\usage{
extract_form_metadata(
  central_url,
  central_email,
  central_password,
  projectID,
  formID,
  form_version = 1,
  isDraft
)
}
\arguments{
\item{central_url}{The url of the ODK central server}

\item{central_email}{The email of your ODK central account}

\item{central_password}{The password to your ODK central account}

\item{projectID}{The ID of the project you are looking at.
To get a list of project, see the
"get_projects" function}

\item{formID}{The ID of the form containing the submissions you are looking at.
To get the list of forms see the "get_forms" function.}

\item{form_version}{The version of the form you are examining. For now
we presume you are looking for the first version of the form}

\item{isDraft}{Whether the form you are submitting to is a draft or not}

\item{submissionID}{The instance ID of the specific submission we want in XML}

\item{xml_string}{The XML string for the data submission you
would like to make}
}
\value{









}
\description{
Extract the metadata for a RHoMIS project
}
\examples{
Get Submissions List

A function for seeing which submissions have been made for
a specific form, from a specific project

Submission All

Get a list of the submissions for a specific form

Get Submission XML

Get the XML of an ODK submission

Submit XML Data

}
