% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NTFP.R
\name{extract_fp_price_and_calorie_conv}
\alias{extract_fp_price_and_calorie_conv}
\title{Extract FP Price and Calorie Values}
\usage{
extract_fp_price_and_calorie_conv(tree_aid_df)
}
\arguments{
\item{tree_aid_df}{Tree Aid Dataset}
}
\description{
Extract FP Price and Calorie Values
}
