% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExtractNewValues.R
\name{extract_livestock_information}
\alias{extract_livestock_information}
\title{Extract Livestock Information}
\usage{
extract_livestock_information(data, crop_name_conversion = pkg.env$crop_name)
}
\arguments{
\item{data}{dataset}

\item{crop_name_conversion}{A tibble to convert crop names to standard values}
}
\value{

}
\description{
Extract livestock names which need to be given
TLU and meat weight conversions
}
