% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redirectModules.R
\name{extract_project_values}
\alias{extract_project_values}
\title{Extract Project values}
\usage{
extract_project_values(rhomis_data)
}
\arguments{
\item{rhomis_data}{A rhomis data set}
}
\description{
Extract all of the values from an individual data set
  
Rpackage file: redirectModules.R
}
