% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-extract-raw-data.R
\name{extract_raw_data_server}
\alias{extract_raw_data_server}
\title{Extract Raw Data}
\usage{
extract_raw_data_server(
  central_url,
  central_email,
  central_password,
  project_name,
  form_name,
  database,
  isDraft
)
}
\arguments{
\item{central_url}{The url of the ODK-central}

\item{central_email}{The email of the ODK-central account being used}

\item{central_password}{The password of the ODK-central account being used}

\item{project_name}{The name of the ODK-central project being processed}

\item{form_name}{The name of the ODK-central form being processed}

\item{database}{The name of the database you would like to save results to}

\item{isDraft}{Whether or not the ODK form you are working with is a draft
or a final version. Only relevant if you are processing a project from ODK central}
}
\description{
This function can be used to extract raw data from
an ODK central server, and save it to a RHoMIS
mongoDB database.
}
