% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UnitsAndConversions.R
\name{extract_secondary_unit_set}
\alias{extract_secondary_unit_set}
\title{Extract Secondary Unit Set}
\usage{
extract_secondary_unit_set(units_and_conversions, name_conversions, unit_type)
}
\arguments{
\item{units_and_conversions}{A list of units conversion tibbles}

\item{name_conversions}{The names for the conversions (e.g. livestock_name)}

\item{unit_type}{The type of unit we are preparing to right to conversion}
}
\description{
In the second processing step
of RHoMIS, there are secondary units
which need to be verified (e.g. livestock_tlu),
this function is used to extract these types
of units and search for existing conversions
within the package (see R/config_package_vars.R "set_secondary_units()").
}
\details{
file: R/UnitsAndConversions.R
}
