% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UnitsAndConversions.R
\name{extract_secondary_units}
\alias{extract_secondary_units}
\title{Extract Secondary Units}
\usage{
extract_secondary_units(units_and_conversions)
}
\arguments{
\item{units_and_conversions}{A list of unit and conversion tibbles}
}
\description{
In the second processing step
of RHoMIS, there are secondary units
which need to be verified (e.g. livestock_tlu),
this function is used to extract these types
of units and search for existing conversions
within the package (see R/config_package_vars.R "set_secondary_units()").
}
\details{
file: R/UnitsAndConversions.R
}
