% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-extract-units.R
\name{extract_units_and_conversions}
\alias{extract_units_and_conversions}
\title{Extract Units and Conversions}
\usage{
extract_units_and_conversions(rhomis_data)
}
\arguments{
\item{rhomis_data}{A "raw" rhomis dataset}
}
\description{
A function to extract units and 
conversion factors from the 
RHoMIS dataset
}
\details{
Rpackage file: 01-extract-units.R
}
