% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UnitsAndConversions.R
\name{extract_units_from_db}
\alias{extract_units_from_db}
\title{Extract Units from database}
\usage{
extract_units_from_db(
  database = "rhomis",
  url = "mongodb://localhost",
  projectID = "core_units",
  formID = "core_units",
  conversion_type,
  collection = "units_and_conversions"
)
}
\arguments{
\item{database}{The name of the database}

\item{url}{The url of the database}

\item{projectID}{The id of the project you are interested in}

\item{formID}{The id of the form you are interested in}

\item{conversion_type}{The type of conversion factor you are extracting}

\item{collection}{The collection where these units are found}
}
\description{
Load all of the units from a local mongoDB
}
\details{
Rpackage file: UnitsAndConversions.R
}
