% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExtractNewValues.R
\name{extract_values_by_project}
\alias{extract_values_by_project}
\title{Extract Values by Project}
\usage{
extract_values_by_project(data)
}
\arguments{
\item{data}{A rhomis data frame}
}
\description{
A function to extract all of the units alongside their
individual form and project IDs.
}
\details{
Rpackage file: ExtractNewValues.R
}
