% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redirectModules.R
\name{extract_values_local}
\alias{extract_values_local}
\title{Extract values}
\usage{
extract_values_local(
  file_path = file_path,
  country_column = "country",
  unique_id_col = "_uuid",
  hh_id_col = NULL,
  id_type = "string",
  proj_id = "KE_ESA_2021",
  form_id = "ESSA"
)
}
\arguments{
\item{file_path}{The path to the raw-data rhomis file}

\item{form_id}{}

\item{overwrite}{Whether}
}
\value{

}
\description{
Extract values
}
