% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fertiliser_amount_to_kg}
\alias{fertiliser_amount_to_kg}
\title{Fertiliser Units}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 2 rows and 2 columns.
}
\usage{
fertiliser_amount_to_kg
}
\description{
A list of common fertiliser amount units from previous RHoMIS datasets
}
\details{
@source NA
}
\keyword{datasets}
