% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-mongodb.R
\name{find_collection}
\alias{find_collection}
\title{Querying whole collection}
\usage{
find_collection(collection, database = "rhomis", url = "mongodb://localhost")
}
\arguments{
\item{collection}{The collection you want to collect the information from}

\item{database}{The database you are querying}

\item{url}{The url of the database you are querying}
}
\value{
A data frame of the data you are querying
}
\description{
Returning all of the information from a mongoDB query in a tabular format
}
\details{
Rpackage file: utils-mongodb.R
}
