% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetaData.R
\name{find_d3_dependencies_network}
\alias{find_d3_dependencies_network}
\title{Find D3 Dependencies Network}
\usage{
find_d3_dependencies_network(
  indicator_name,
  list_of_indicators,
  d3_list = list()
)
}
\arguments{
\item{indicator_name}{The name of the indicator}

\item{list_of_indicators}{The entire list of RHoMIS indicators}

\item{d3_list}{An existing d3 formatted list}
}
\description{
Find indicator nested dependencies and
return them in a format
which can be used to draw a d3 network plot
}
