% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UnitsAndConversions.R
\name{find_db_units}
\alias{find_db_units}
\title{Find db Units}
\usage{
find_db_units(projectID, formID, url, collection, database)
}
\arguments{
\item{projectID}{The name of the project}

\item{formID}{The name of the form}

\item{url}{The URL of the mongoDB you are querying}

\item{collection}{The collection storing administrative project information}

\item{database}{The name of the database containing the information}
}
\description{
The names of the units stored in the RHoMIS data base
are stored within the "projectData" collection. This
function lets you query which types of units were found
for a particular project
}
\details{
Rpackage file: UnitsAndConversions.R
}
