% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExtractNewValues.R
\name{find_loop_number_and_extract_values}
\alias{find_loop_number_and_extract_values}
\title{Find the Number of Loops and Extract new values}
\usage{
find_loop_number_and_extract_values(data, column_pattern)
}
\arguments{
\item{data}{The dataset to extract the new values from}

\item{column_pattern}{The repeat column pattern e.g. "crop_repeat"}
}
\description{
Find the number of loops for a looped RHoMIS data
set.
}
\details{
Rpackage file: ExtractNewValues.R
}
