% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetaData.R
\name{find_nested_dependencies_network}
\alias{find_nested_dependencies_network}
\title{Find Nested Dependencies}
\usage{
find_nested_dependencies_network(
  indicator_name,
  indicator_list = indicator_list
)
}
\arguments{
\item{indicator_name}{The name of the indicator}

\item{indicator_list}{The list of indicators}
}
\description{
In the RHoMIS datasets, there are indicators
which depend on indicators, indicators which depend
on individual columns, and indicators which depend on columns
embedded in loops. It is essential that we know how every
indicator is linked in the data so that we can understand
issues in survey design. This function tries to achieve this.
}
