% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReformatLoops.R
\name{find_unique_case}
\alias{find_unique_case}
\title{Find Unique Names}
\usage{
find_unique_case(data, name_column)
}
\arguments{
\item{data}{The data-frame containing the loops of concern}

\item{name_column}{The original name of the loop (e.g. "crop_name")}

\item{number_of_loops}{The number of loops for the variable of concern}
}
\value{
A list of all of the unique entries for the name column
}
\description{
A function to find all of the unique names for a particular looped variable (e.g. livestock, crops, off-farm incomes)
}
\examples{
name_column <- "crop_name"
data <- as_tibble(list(crop_name_1=c("banana", "cassava", NA, "millet"),
                       crop_name_2=c("cassava",NA, "melon", "maize"),
                       random_crop_name_2=c("blue", "green",  "red",NA),
                       crop_name=c("orange", "purple", NA, "black")))
expected_result <- c("banana","cassava","millet","melon","maize")
find_unique_case(data, name_column)
}
