% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExtractNewValues.R
\name{find_unique_values}
\alias{find_unique_values}
\title{Finding Unique Values}
\usage{
find_unique_values(data)
}
\arguments{
\item{data}{A tibble for which we hope to find all of the unique values}
}
\value{
A list of the unique values, NAs excluded
}
\description{
A function to find the unique values for all of a given tibble.
}
\details{
Rpackage file: ExtractNewValues.R
}
\examples{
data <- tibble::tibble(
  crop_name_1 = c("maize", "cassava", "millet"),
  crop_name_2 = c("potato", NA, "cucumber")
)
# Expect the result c("maize", "cassava", "millet","potato", "cucumber")
}
