% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FoodSecurity.R
\name{food_security_calculations}
\alias{food_security_calculations}
\title{Food Security Calculations All}
\usage{
food_security_calculations(data)
}
\arguments{
\item{data}{Survey data containing either FIES or
HFIAS data}
}
\description{
Calculate food security calculations
based on whether HFIAS or FIES data were collected in the RHoMIS
survey
}
\details{
Rpackage file: FoodSecurity.R
}
