% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fp_amount_to_kg}
\alias{fp_amount_to_kg}
\title{FP amount Conversions}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 4 rows and 2 columns.
}
\source{
NA
}
\usage{
fp_amount_to_kg
}
\description{
Conversions of FP amounts
to KG
}
\keyword{datasets}
