% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NTFP.R
\name{fp_income_calculations}
\alias{fp_income_calculations}
\title{NTFP Income Calculation}
\usage{
fp_income_calculations(
  rhomis_data,
  unit_conv_tibble = NULL,
  fp_sold_kg_per_year_column,
  fp_sold_units_column,
  fp_sold_income_column,
  new_fp_sold_income,
  product_type
)
}
\arguments{
\item{product_type}{}
}
\description{
Calculate the Income from NTFPs
in LCU per year
}
