% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fp_income_per_freq_to_lcu_per_year}
\alias{fp_income_per_freq_to_lcu_per_year}
\title{FP Income pe Year}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 4 rows and 2 columns.
}
\source{
NA
}
\usage{
fp_income_per_freq_to_lcu_per_year
}
\description{
Conversions of FP incomes
to lcu per year
}
\keyword{datasets}
