% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fp_products}
\alias{fp_products}
\title{FP Products}
\format{
An object of class \code{list} of length 8.
}
\source{
NA
}
\usage{
fp_products
}
\description{
The NTFP products which
are contained within the
NTFP module
}
\keyword{datasets}
