% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{function_list}
\alias{function_list}
\title{Function List}
\format{
An object of class \code{list} of length 33.
}
\source{
NA
}
\usage{
function_list
}
\description{
A list of the functions used to
calculate various RHoMIS indicators
}
\keyword{datasets}
