% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenderControlAll.R
\name{gender_control_summary}
\alias{gender_control_summary}
\title{Gender Control Summary}
\usage{
gender_control_summary(
  processed_data,
  indicator_data,
  gender_categories = pkg.env$gender_categories
)
}
\arguments{
\item{processed_data}{A processed RHoMIS dataset}

\item{indicator_data}{An indicator dataset, where you would like to append the result}

\item{gender_categories}{A vector of the gender categories in the dataset.}
}
\value{

}
\description{
Generate a summary of all of the gender controlled
indicators
}
