% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TotalIncomes.R
\name{gendered_off_farm_income_indicator}
\alias{gendered_off_farm_income_indicator}
\title{Gendered Off-Farm Incomes Indicator}
\usage{
gendered_off_farm_income_indicator(data)
}
\arguments{
\item{data}{RHoMIS data including off-farm income loops}
}
\value{

}
\description{
Gendered Off-Farm Incomes Indicator
}
