% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenerateMockResponses.R
\name{generate_mock_response}
\alias{generate_mock_response}
\title{Generate Mock Response}
\usage{
generate_mock_response(survey_path)
}
\arguments{
\item{survey_path}{The path to the xls survey file}
}
\value{

}
\description{
This is a function to generate a mock response based on
the survey file alone
}
