% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenerateMockResponses.R
\name{generate_random_row}
\alias{generate_random_row}
\title{Generate a Random Row}
\usage{
generate_random_row(
  submission_xml,
  survey,
  choices,
  survey_row,
  xml_level,
  group_at_level
)
}
\arguments{
\item{submission_xml}{The previous xml which must be
appended}

\item{survey}{The survey sheet to read}

\item{choices}{The choices to read}

\item{survey_row}{The index (which row) of the
survey sheet which we must examine}

\item{xml_level}{The level at which we are looking
i.e, how many tabs to put at the start of this line}

\item{group_at_level}{A dictionary telling us which
"names" are associate with level of indentation, making
it easier to close the groups when the survey groups
when they have ended.}
}
\value{

}
\description{
For each row in the survey file generate
the xml row needed for submission.
}
