% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenerateMockResponses.R
\name{generate_tab_spaces}
\alias{generate_tab_spaces}
\title{Generate Tab Spaces}
\usage{
generate_tab_spaces(form_level)
}
\arguments{
\item{form_level}{The level (amount of indentation)
we need for a particular line in the xml response}
}
\description{
Make the necessary number of
spaces based on the level of the xml form
}
\details{
Rpackage file: GenerateMockResponses.R
}
