% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConnectToCentral.R
\name{get_email_token}
\alias{get_email_token}
\title{Get Email Token}
\usage{
get_email_token(central_url, central_email, central_password)
}
\arguments{
\item{central_url}{The url which points to an ODK Central sever}

\item{central_email}{The email linked to your ODK central account}

\item{central_password}{The password used for your central account}
}
\value{
This function will return a token. This token is useful for
}
\description{
A function used to get the email token required to connext to ODK central
}
\details{
Rpackage file: ConnectToCentral.R
}
