% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConnectToCentral.R
\name{get_forms}
\alias{get_forms}
\title{Get forms}
\usage{
get_forms(central_url, central_email, central_password, projectID)
}
\arguments{
\item{central_url}{The url of the ODK central server}

\item{central_email}{The email of your ODK central account}

\item{central_password}{The password to your ODK central account}

\item{projectID}{The ID of the project you are looking at.
To get a list of project, see the
"get_projects" function}
}
\value{
A table of the forms available for a specific project
}
\description{
Get the forms for a specific ODK central project
}
