% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HouseholdSize.R
\name{get_household_size_conversion}
\alias{get_household_size_conversion}
\title{Get household Size Conversions}
\usage{
get_household_size_conversion()
}
\value{

}
\description{
Get household Size Conversions
}
