% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConnectToSurveyBuilder.R
\name{get_individual_survey_builder_project}
\alias{get_individual_survey_builder_project}
\title{Get Individual Survey Builder Project}
\usage{
get_individual_survey_builder_project(
  survey_builder_url,
  survey_builder_access_token,
  projectID
)
}
\arguments{
\item{survey_builder_url}{The URL of the survey builder application}

\item{survey_builder_access_token}{The access toke to access the
survey builder API}

\item{projectID}{The ID of the project you want to get
individual information for}
}
\value{

}
\description{
The survey builder contains meta-data about the people
using RHoMIS 2.0. This function connects to the survey
builder API to get the meta-data information from the
individual projects
}
