% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConnectToCentral.R
\name{get_project_id_from_name}
\alias{get_project_id_from_name}
\title{Get Project ID from name}
\usage{
get_project_id_from_name(
  project_name,
  central_url,
  central_email,
  central_password
)
}
\arguments{
\item{project_name}{The name of the project from which you require the ID}

\item{central_url}{The url which points to an ODK Central sever}

\item{central_email}{The email linked to your ODK central account}

\item{central_password}{The password used for your central account}
}
\value{

}
\description{
Get Project ID from name
}
