% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConnectToCentral.R
\name{get_projects}
\alias{get_projects}
\title{Get Projects}
\usage{
get_projects(central_url, central_email, central_password)
}
\arguments{
\item{central_url}{The url of the ODK central server}

\item{central_email}{The email of your ODK central account}

\item{central_password}{The password to your ODK central account}
}
\value{
A table of all of the projects and their details
}
\description{
Get a table of all of the projects on the ODK central server
}
\details{
Rpackage file: ConnectToCentral.R
}
\examples{
central_url <- "https://central.rhomis.cgiar.org"
central_email <- "my_email"
central_password <- "my_password"
# get_projects(central_url,central_email,central_password)
}
