% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-calculate-prices.R
\name{get_secondary_conversions}
\alias{get_secondary_conversions}
\title{Get Secondary Conversions}
\usage{
get_secondary_conversions(
  rhomis_data,
  units_and_conversions,
  gender_categories = pkg.env$gender_categories
)
}
\arguments{
\item{rhomis_data}{A "raw" rhomis dataset}
}
\description{
Rpackage file: 02-calculate-prices.R
}
