% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConnectToCentral.R
\name{get_submissions_list}
\alias{get_submissions_list}
\title{Get Submissions List}
\usage{
get_submissions_list(
  central_url,
  central_email,
  central_password,
  projectID,
  formID
)
}
\arguments{
\item{central_url}{The url of the ODK central server}

\item{central_email}{The email of your ODK central account}

\item{central_password}{The password to your ODK central account}

\item{projectID}{The ID of the project you are looking at. To get a list of project, see the
"get_projects" function}

\item{formID}{The ID of the form containing the submissions you are looking at.
To get the list of forms see the "get_forms" function.}
}
\value{

}
\description{
A function for seeing which submissions have been made for
a specific form, from a specific project
}
