% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConnectToSurveyBuilder.R
\name{get_survey_builder_projects}
\alias{get_survey_builder_projects}
\title{Get Survey Builder Project}
\usage{
get_survey_builder_projects(survey_builder_url, survey_builder_access_token)
}
\arguments{
\item{survey_builder_url}{The URL of the survey builder application}

\item{survey_builder_access_token}{The access toke to access the
survey builder API}
}
\description{
The survey builder contains meta-data about the people
using RHoMIS 2.0. This function connects to the survey
builder API to get the meta-data information from
multiple projects
}
\details{
Rpackage file: ConnectToSurveyBuilder.R
}
