% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConnectToCentral.R
\name{get_users}
\alias{get_users}
\title{Get Users}
\usage{
get_users(central_url, central_email, central_password)
}
\arguments{
\item{central_url}{The url of the ODK central server}

\item{central_email}{The email of your ODK central account}

\item{central_password}{The password to your ODK central account}
}
\value{
A tibble of all of the users in the central database
}
\description{
Get a table of users from the ODK central database
}
