% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-odk-central.R
\name{get_xls_survey_file}
\alias{get_xls_survey_file}
\title{Get xls survey file}
\usage{
get_xls_survey_file(
  central_url,
  central_email,
  central_password,
  projectID,
  formID,
  file_destination = NULL,
  form_version = 1
)
}
\arguments{
\item{central_url}{The url of the ODK central server}

\item{central_email}{The email of your ODK central account}

\item{central_password}{The password to your ODK central account}

\item{projectID}{The ID of the project you are looking at.
To get a list of project, see the
"get_projects" function}

\item{formID}{The XML form ID from a specific project}

\item{file_destination}{Allow the user to specify the destination of
the xls file they are working with.
include the file name and the extension. Must make sure the directory exists}

\item{form_version}{The version of the form you are examining. For now
we presume you are looking for the first version of the form}
}
\description{
A function for getting the xls survey from odk central,
writing it to a file, and returning the file path
}
\details{
Rpackage file: utils-odk-central.R
}
