% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConnectToCentral.R
\name{get_xml_form_id_from_name}
\alias{get_xml_form_id_from_name}
\title{Get Form ID from Name}
\usage{
get_xml_form_id_from_name(
  form_name,
  projectID,
  central_url,
  central_email,
  central_password
)
}
\arguments{
\item{form_name}{The name of the form for which you would like the ID}

\item{projectID}{The ID for the project you are searching for}

\item{central_url}{The url which points to an ODK Central sever}

\item{central_email}{The email linked to your ODK central account}

\item{central_password}{The password used for your central account}
}
\value{

}
\description{
Sometimes it is necessary to get the form
ID from the project name, in particular
}
