% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hdds_10_test}
\alias{hdds_10_test}
\title{HDDS 10 test}
\format{
A tibble
}
\source{
A sample rhomis dataset
}
\usage{
hdds_10_test
}
\description{
A dataset used for testing dietary diversity calculations with 10 food groups
}
\keyword{datasets}
