% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hdds_14_test}
\alias{hdds_14_test}
\title{HDDS 14 test}
\format{
A tibble
}
\source{
A sample rhomis dataset
}
\usage{
hdds_14_test
}
\description{
A dataset used for testing dietary diversity calculations with 14 food groups
}
\keyword{datasets}
