% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LivestockCalculations.R
\name{honey_amount_calculation}
\alias{honey_amount_calculation}
\title{Honey Amount Calculation}
\usage{
honey_amount_calculation(data, unit_conv_tibble = NULL)
}
\arguments{
\item{data}{The data containing livestock loops}

\item{unit_conv_tibble}{A tibble of honey amount conversions}
}
\value{

}
\description{
Calculating the amount of honey produced from RHoMIS data
}
