% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LivestockCalculations.R
\name{honey_amount_sold_and_consumed_calculations}
\alias{honey_amount_sold_and_consumed_calculations}
\title{Honey Sold and Consumed Calculations}
\usage{
honey_amount_sold_and_consumed_calculations(data)
}
\arguments{
\item{data}{Data containing RHoMIS
livestock loops}
}
\description{
Calculating the amounts of honey sold and consumed
  
Rpackage file: LivestockCalculations.R
}
