% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{honey_amount_to_l}
\alias{honey_amount_to_l}
\title{Honey Amount Units}
\format{
A tibble
}
\source{
NA
}
\usage{
honey_amount_to_l
}
\description{
Conversion factors for volume of honey into
kilograms
}
\keyword{datasets}
