% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{honey_calories}
\alias{honey_calories}
\title{Honey Calories}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 1 rows and 2 columns.
}
\source{
NA
}
\usage{
honey_calories
}
\description{
Conversion factor for calories from honey
}
\keyword{datasets}
