% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HouseholdSize.R
\name{household_roster_to_wide}
\alias{household_roster_to_wide}
\title{Household Roster to Wide}
\usage{
household_roster_to_wide(data)
}
\arguments{
\item{data}{RHoMIS data with household Roster information}
}
\description{
Moving household roster size to wide
 
Rpackage file: HouseholdSize.R
}
