% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LivestockCalculations.R
\name{identify_number_of_heads_for_livestock_loops}
\alias{identify_number_of_heads_for_livestock_loops}
\title{Identify Number of Livestock Heads in Loops}
\usage{
identify_number_of_heads_for_livestock_loops(
  livestock_name_column,
  livestock_heads_df
)
}
\arguments{
\item{livestock_name_column}{A vector of livestock names}

\item{livestock_heads_df}{A tibble of livestock head nunbers}
}
\description{
Livestock heads information is collected seperately
the livestock loops. Howver in some cases, for example in
the eggs_amount calculations, it is useful to know the livestock heads
during the loops. This function allows us to do so
  
Rpackage file: LivestockCalculations.R
}
