% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HouseholdSize.R
\name{identify_person_category}
\alias{identify_person_category}
\title{Identify Person Category}
\usage{
identify_person_category(age, gender)
}
\arguments{
\item{age}{The age vector for the specific loop}

\item{gender}{The gender vector for the specific loop}
}
\description{
For the household roster, categories
each person based on the MAE conversion information
 
Rpackage file: HouseholdSize.R
}
