% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CleanColumnNames.R
\name{identify_repeat_columns}
\alias{identify_repeat_columns}
\title{Identify Repeat Columns}
\usage{
identify_repeat_columns(column_names)
}
\arguments{
\item{column_names}{The list of column names}
}
\description{
Identify repeat loop columns in a rhomis
dataset
}
