% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenerateMockResponses.R
\name{identify_repeat_locations}
\alias{identify_repeat_locations}
\title{Identify Repeat Locations}
\usage{
identify_repeat_locations(survey)
}
\arguments{
\item{survey}{The survey sheet of the
excel file as a tibble}
}
\value{

}
\description{
Find the rows in the survey file
where questions are in repeat loops
}
