% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{indicator_list}
\alias{indicator_list}
\title{Indicator List}
\format{
An object of class \code{list} of length 161.

An object of class \code{list} of length 161.
}
\source{
NA

NA
}
\usage{
indicator_list

indicator_list
}
\description{
A list of the different
indicators calculated in this
package, along with the functions
used to calculate them

A list of the different
indicators calculated in this
package, along with the functions
used to calculate them
}
\keyword{datasets}
