% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetaData.R
\name{indicator_mapping_to_df}
\alias{indicator_mapping_to_df}
\title{Indicator Mapping to DF}
\usage{
indicator_mapping_to_df()
}
\description{
Convert the dependent indicator mapping
structure to a data frame
}
