% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReformatLoops.R
\name{insert_gender_columns_in_core_data}
\alias{insert_gender_columns_in_core_data}
\title{Insert Gender Columns in Core Data}
\usage{
insert_gender_columns_in_core_data(
  data,
  original_column,
  control_column,
  loop_structure = F,
  gender_control_categories = pkg.env$gender_categories
)
}
\arguments{
\item{data}{The original data set containing the numeric data
and the gender control column used to split it}

\item{original_column}{The original numeric column which
needs to be split by gender}

\item{control_column}{The column indicating the groups which are
controlling a resource. Must include the values "male_adult","female_adult" etc...}

\item{loop_structure}{Indicating whether or not the controlled
resource is located within a loop structure. So far all gender control
columns are located within this looping structure}

\item{gender_control_categories}{The gender control categories included in the survey.}
}
\value{

}
\description{
Splitting a particular column in the RHoMIS
data by the genders groups which control that resource
}
