% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{land_area_units}
\alias{land_area_units}
\title{Land Area Units}
\format{
A tibble
}
\source{
NA
}
\usage{
land_area_units
}
\description{
Conversion of land area into hectares
}
\keyword{datasets}
