% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LandSize.R
\name{land_size_calculation}
\alias{land_size_calculation}
\title{Land Size Calculation}
\usage{
land_size_calculation(
  data,
  units = land_area_units$units,
  unit_conversions = land_area_units$conversions
)
}
\arguments{
\item{data}{RHoMIS data}

\item{units}{Land size units}

\item{unit_conversions}{Unit conversions for the land size units}
}
\value{

}
\description{
Calculate land sizes in hectares using
RHoMIS data
}
