% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LivestockCalculations.R
\name{livestock_calculations_all}
\alias{livestock_calculations_all}
\title{Livestock calculations all}
\usage{
livestock_calculations_all(
  data,
  livestock_weights_names = livestock_weights$animal,
  livestock_weights_conversions = livestock_weights$weight_kg,
  eggs_amount_units_all = eggs_amount_units$unit,
  eggs_amount_unit_conversions_all = eggs_amount_units$conversion_factor,
  eggs_price_time_units_all = eggs_price_time_units$unit,
  eggs_price_time_unit_conversions_all = eggs_price_time_units$conversion_factor,
  honey_amount_units_all = honey_amount_units$units,
  honey_amount_unit_conversions_all = honey_amount_units$conversion_factors,
  milk_amount_units_all = milk_amount_units$unit,
  milk_amount_unit_conversions_all = milk_amount_units$conversion_factor,
  milk_price_time_units_all = milk_price_time_units$unit,
  milk_price_time_unit_conversions_all = milk_price_time_units$conversion_factor
)
}
\arguments{
\item{data}{RHoMIS data, containing livestock loops}

\item{livestock_weights_names}{The names of livestock}

\item{livestock_weights_conversions}{Conversion factors to convert livestock
names into kg of meat}

\item{eggs_amount_units_all}{Units for eggs harvested}

\item{eggs_amount_unit_conversions_all}{Conversion factors to convert the
eggs harvested units into a per year value}

\item{eggs_price_time_units_all}{Egg price units in a unit of time}

\item{eggs_price_time_unit_conversions_all}{Conversion factors to convert
the price units into a yearly income}

\item{honey_amount_units_all}{Units for amounts of honey
collected}

\item{honey_amount_unit_conversions_all}{Conversion factors for
the honey amount units}

\item{milk_amount_units_all}{Milk amount units}

\item{milk_amount_unit_conversions_all}{Conversion factors for the
milk amount units}

\item{milk_price_time_units_all}{Units for milk price}

\item{milk_price_time_unit_conversions_all}{Conversion factors for the milk
price units}
}
\value{

}
\description{
Carrying out all calculations on RHoMIS
livestock loops.
}
