% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CropAndLivestockDiversity.R
\name{livestock_diversity}
\alias{livestock_diversity}
\title{Livestock Diversity}
\usage{
livestock_diversity(data, indicator_data)
}
\arguments{
\item{data}{RHoMIS dataset}

\item{indicator_data}{A tibble to store indicators, if crop diversity can be calculated, then the indicator will be appended.}
}
\description{
Calculating the number of livestock kept by a household
}
\details{
Rpackage file: CropAndLivestockDiversity.R
}
