% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{livestock_name}
\alias{livestock_name}
\title{Livestock Name}
\format{
An object of class \code{character} of length 12.
}
\usage{
livestock_name
}
\description{
A list of common names found in previous RHoMIS datasets
}
\details{
@source NA
}
\keyword{datasets}
