% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LivestockCalculations.R
\name{livestock_tlu_calculations}
\alias{livestock_tlu_calculations}
\title{Livestock TLU Calculations}
\usage{
livestock_tlu_calculations(
  data,
  livestock_name_conversion_tibble,
  livestock_tlu_conversions
)
}
\arguments{
\item{data}{RHoMIS processed dataset}

\item{livestock_name_conversion_tibble}{A tibble
of conversions for livestock names}

\item{livestock_tlu_conversions}{A tibble for converting 
livestock names into TLUss}
}
\description{
Calculate amount of livestock kept in
tropical livestock units (TLU).
  
Rpackage file: LivestockCalculations.R
}
