% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{livestock_weight_kg}
\alias{livestock_weight_kg}
\title{Livestock Weights}
\format{
A tibble
}
\source{
NA
}
\usage{
livestock_weight_kg
}
\description{
Conversion factors for the weights
of livestock when they are killed for meat
}
\keyword{datasets}
