% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UnitsAndConversions.R
\name{load_all_db_units}
\alias{load_all_db_units}
\title{Load all db units}
\usage{
load_all_db_units(
  unit_list,
  database = "rhomis",
  projectID = "core_units",
  formID = "core_units"
)
}
\arguments{
\item{unit_list}{The list of units which are to be queried and loaded into the global environment}

\item{database}{The database you are querying units from}

\item{projectID}{The Name of the project}

\item{formID}{The name of the form}
}
\value{

}
\description{
Load all units from a local MongoDB databases
}
\examples{
loop over the possible list of unit conversion csv file names
evaluate the string denoting the variable name to be used
make dummy tibble
}
