% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UnitsAndConversions.R
\name{load_calorie_conversions}
\alias{load_calorie_conversions}
\title{Load Calorie Conversions}
\usage{
load_calorie_conversions(base_folder, id_rhomis_dataset)
}
\arguments{
\item{base_folder}{The path to the folder containing the units to load}

\item{id_rhomis_dataset}{A vector including the ID of the RHoMIS datasets being processed}
}
\description{
Load units for a particular project from csv and
load them into the global environment
}
\details{
Rpackage file: UnitsAndConversions.R
}
