% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunAll.R
\name{load_calorie_csvs}
\alias{load_calorie_csvs}
\title{Load Calorie CSVs}
\usage{
load_calorie_csvs(folder, ids_rhomis_dataset)
}
\arguments{
\item{folder}{The folder containing all of the calorie conversions.}

\item{id_rhomis_dataset}{A vector of the RHoMIS IDs}
}
\value{

}
\description{
Load Calorie CSVs
}
