% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UnitsAndConversions.R
\name{load_local_units}
\alias{load_local_units}
\title{Load local units}
\usage{
load_local_units(units_folder, id_rhomis_dataset, unit_type = "primary")
}
\arguments{
\item{units_folder}{The path to the folder containing the units to load}

\item{id_rhomis_dataset}{A vector including the ID of the RHoMIS datasets being processed}

\item{unit_type}{Whether you are reading "primary" or "secondary" units.}
}
\description{
Load units for a particular project from csv and
load them into the global environment
}
\details{
Rpackage file: UnitsAndConversions.R
}
