% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunAll.R
\name{load_rhomis_csv}
\alias{load_rhomis_csv}
\title{Load RHoMIS CSV}
\usage{
load_rhomis_csv(
  file_path,
  country_column = "country",
  unique_id_col = "_uuid",
  hh_id_col = NULL,
  id_type = "string",
  proj_id = NULL,
  form_id = NULL,
  proj_id_col = NULL,
  form_id_col = NULL,
  overwrite = FALSE,
  repeat_column_names = c("crop_repeat", "livestock_repeat", "offfarm_repeat",
    "offfarm_income_repeat", "hh_pop_repeat", "hh_rep")
)
}
\arguments{
\item{file_path}{The filepath of the RHoMIS csv}

\item{country_column}{The name of the column containing the country}

\item{id_type}{Indicator of whether you are providing a single ID}

\item{proj_id}{A single string to be used as the project ID for all households}

\item{form_id}{A single string to be used as the form ID for all households}

\item{proj_id_col}{The name of the column containing the project IDs}

\item{form_id_col}{The name of the column containing the form IDs}

\item{overwrite}{True if you would like to overwrite previous ID column, false if would not like to overwrite existing IDs}

\item{repeat_column_names}{The types of repeat column name}
}
\value{
A tibble of RHoMIS data
}
\description{
Load a Raw RHoMIS csv file, collected using ODK, and
convert the column names into a shortened, standardised
version.
}
\examples{

}
