% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-csv-data.R
\name{load_survey_csv}
\alias{load_survey_csv}
\title{Load Survey CSV}
\usage{
load_survey_csv(
  file_path,
  country_column = "country",
  unique_id_col = "_uuid",
  hh_id_col = NULL,
  id_type = c("string", "column"),
  proj_id = NULL,
  form_id = NULL,
  overwrite = FALSE,
  repeat_columns = pkg.env$repeat_columns
)
}
\arguments{
\item{file_path}{The filepath of the RHoMIS csv}

\item{country_column}{The name of the column containing the country}

\item{unique_id_col}{The column in the dataset which contains unique IDs (usually _uuid)}

\item{hh_id_col}{The column containing household IDs}

\item{id_type}{Indicator of whether you are providing a single ID}

\item{proj_id}{Either a single string to be used as the project ID for all households, or the name of the column containing the project IDs (depending on id_type)}

\item{form_id}{Either a single string to be used as the form ID for all households, or the name of the column containing the form IDs (depending on id_type)}

\item{overwrite}{True if you would like to overwrite previous ID column, false if would not like to overwrite existing IDs}

\item{repeat_columns}{The types of repeat column name}
}
\value{
A tibble of RHoMIS data
}
\description{
Load a Raw or processed RHoMIS survey csv file, collected using ODK, 
and convert the column names into a shortened, standardised
version. This function can either be used to read
 
Rpackage file: utils-csv-data.R
}
