% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AuthenticateUser.R
\name{login}
\alias{login}
\title{Login}
\usage{
login(url, email, password)
}
\arguments{
\item{url}{The URL of the authentication server}

\item{email}{The email of the account being authenticated}

\item{password}{The password of the account being authenticated}
}
\value{
A token
}
\description{
A function for logging in to the RHoMIS 2.0
token based authenticator
}
