% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReformatLoops.R
\name{loop_to_column_conversion}
\alias{loop_to_column_conversion}
\title{Item Number to Column Row Conversion}
\usage{
loop_to_column_conversion(
  data,
  name_column,
  variable_to_convert,
  number_of_loops,
  unique_names
)
}
\arguments{
\item{data}{The data containing the loops you need to convert}

\item{name_column}{The column containing the variable which will eventually
column header (see example)}

\item{unique_names}{}
}
\value{

}
\description{
Converting from loop to a format that is easy to analyze.
}
