% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunAll.R
\name{make_id_columns}
\alias{make_id_columns}
\title{Make ID Columns}
\usage{
make_id_columns(
  data,
  country_column = "country",
  unique_id_col = "_uuid",
  hh_id_col = NULL,
  id_type = "string",
  proj_id = NULL,
  form_id = NULL,
  proj_id_col = NULL,
  form_id_col = NULL
)
}
\arguments{
\item{data}{The rhomis data set as a tibble}

\item{country_column}{The name of the column containing the country string, as collected in the survey}

\item{unique_id_col}{The name of the column containing unique id record. This is produced by the server accepting ODK records}

\item{hh_id_col}{The household ID}

\item{id_type}{The type of ID you would like to enter for projects and forms. If you select "string", then fill in the proj_id and form_id arguments, with the project id and form id you would like to use. If selecting "column", enter the name of the column (proj_id_col) containing the project ID you would like to use, and the name of the column (form_id_col) containing the form ids you would like to use.}

\item{proj_id}{The name or ID of the project (string), see id_type varaible}

\item{form_id}{The name or ID of the form (string), see id_type variable}

\item{proj_id_col}{The name of the column containing the project id, see id_type variable}

\item{form_id_col}{The name of the column containing the form id, see id_type variable}
}
\value{

}
\description{
Make ID columns for form, project, and household
}
