% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redirectModules.R
\name{make_new_dataset}
\alias{make_new_dataset}
\title{Make new dataset}
\usage{
make_new_dataset(rhomis_data)
}
\arguments{
\item{rhomis_data}{A rhomis_dataset including IDs}
}
\description{
Rpackage file: redirectModules.R
}
