% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CleanRawData.R
\name{make_per_project_conversion_tibble}
\alias{make_per_project_conversion_tibble}
\title{Make Per Project conversion tibble}
\usage{
make_per_project_conversion_tibble(proj_id_vector, unit_conv_tibble)
}
\arguments{
\item{proj_id_vector}{A vector of project IDs}

\item{unit_conv_tibble}{A conversion table containing columns "survey_value" and "conversion_factor"}
}
\value{

}
\description{
If no conversion table is provided, construct
a conversion tibble, including project IDs, from the
the in built conversion tables
}
