% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReformatLoops.R
\name{map_to_wide_format}
\alias{map_to_wide_format}
\title{Map To Wide Format}
\usage{
map_to_wide_format(data, name_column, number_of_loops, column_prefixes)
}
\arguments{
\item{data}{The original dataframe containing the data to be reformatted}

\item{name_column}{The name column (e.g. crop_name_1)}

\item{number_of_loops}{The number of loops for this particular parameter}

\item{other_column_prefixes}{The other columns which need to be exported to a wider named format}
}
\value{
a list of data frames for each of the variables
}
\description{
A function to convert RHoMIS loops into a wider named format
}
