% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{meat_calories}
\alias{meat_calories}
\title{Meat Calories}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 12 rows and 2 columns.
}
\source{
NA
}
\usage{
meat_calories
}
\description{
Calorie conversion factors for meat
}
\keyword{datasets}
