% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LivestockCalculations.R
\name{meat_prices}
\alias{meat_prices}
\title{Meat price per kg}
\usage{
meat_prices(data)
}
\arguments{
\item{data}{RHoMIS data including columns "meat_sold_kg_per_year" and "meat_sold_income".}
}
\description{
Calculating the price of meat which was sold.
}
\details{
Rpackage file: LivestockCalculations.R
}
