% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LivestockCalculations.R
\name{meat_sold_and_consumed_calculation}
\alias{meat_sold_and_consumed_calculation}
\title{Meat Sold and Consumed Calculation}
\usage{
meat_sold_and_consumed_calculation(data)
}
\arguments{
\item{data}{Data containing "livestock_name", "meat_kg_per_year",
"meat_sold_props_numeric","meat_consumed_props_numeric".}
}
\value{

}
\description{
Calculate the amount of meat sold and consumed in kg.
Note this only works if you have calculated the meat collected
in kg.
}
