% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LivestockCalculations.R
\name{meat_uses}
\alias{meat_uses}
\title{Meat Use Calculations}
\usage{
meat_uses(data)
}
\arguments{
\item{data}{The RHoMIS data containing livestock loops}
}
\description{
Calculating the numeric proportions of meat used for
eating and selling
  
Rpackage file: LivestockCalculations.R
}
