% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataReformatting.R
\name{merge_original_and_other_unit_single_pair}
\alias{merge_original_and_other_unit_single_pair}
\title{Merge original and other units pair}
\usage{
merge_original_and_other_unit_single_pair(original_column, other_column)
}
\arguments{
\item{original_column}{the original column you would like to change}

\item{other_column}{the other values that need to be brought into the main column}
}
\description{
Many of the RHoMIS variables come from multiple
select, with the option to specify "other" by
free-text entry. This function allows you to bring
the "other" category into the main column, for a single pair of columns.
}
\details{
Rpackage file: DataReformatting.R
}
