% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataReformatting.R
\name{merge_original_and_other_units}
\alias{merge_original_and_other_units}
\title{Merge original and other units}
\usage{
merge_original_and_other_units(data, main_column, other_column, loop_structure)
}
\arguments{
\item{data}{The data you would like to convert}

\item{main_column}{The main column you would like to add the "other" options}

\item{other_column}{The name of the "other" column where you would like to
integrate your values}

\item{loop_structure}{Whether or not the "other" columns you are converting come
in a loop, or whether it is simple a pair of columns}
}
\value{

}
\description{
Many of the RHoMIS variables come from multiple
select, with the option to specify "other" by
free-text entry. This function allows you to bring
the "other" category into the main column, facilitating analysis.
}
\examples{

}
