% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LivestockCalculations.R
\name{milk_amount_calculations}
\alias{milk_amount_calculations}
\title{Milk amount calculations}
\usage{
milk_amount_calculations(
  data,
  milk_units = milk_amount_units$unit,
  milk_unit_conversions = milk_amount_units$conversion_factor
)
}
\arguments{
\item{data}{The dataset containing livestock loops for RHoMIS}

\item{milk_units}{The units for milk amounts}

\item{milk_unit_conversions}{The conversions for each of these units to litres per year}
}
\value{

}
\description{
Calculating the amount of milk collected each year
}
