% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{milk_amount_to_l}
\alias{milk_amount_to_l}
\title{Milk Amount Units}
\format{
A tibble

A tibble
}
\source{
NA

NA
}
\usage{
milk_amount_to_l

milk_amount_to_l
}
\description{
Conversion factors for the amounts
of milk collected

Conversion factors for the price of
milk, can both be a time or a per
unit volume
}
\keyword{datasets}
