% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{milk_amount_units}
\alias{milk_amount_units}
\title{Milk Amount Units}
\format{
A tibble
}
\source{
NA
}
\usage{
milk_amount_units
}
\description{
Conversion factors for the amounts
of milk collected
}
\keyword{datasets}
