% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{milk_calories}
\alias{milk_calories}
\title{Milk Calories}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 12 rows and 2 columns.
}
\source{
NA
}
\usage{
milk_calories
}
\description{
Calorie conversion factors for milk
}
\keyword{datasets}
