% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LivestockCalculations.R
\name{milk_income_calculations}
\alias{milk_income_calculations}
\title{Milk Income Calculations
  
Rpackage file: LivestockCalculations.R}
\usage{
milk_income_calculations(data, unit_conv_tibble = NULL)
}
\arguments{
\item{data}{Dataset containing all milk income information}

\item{unit_conv_tibble}{A tibble containing all of the milk income conversion factors}
}
\description{
Milk Income Calculations
  
Rpackage file: LivestockCalculations.R
}
