% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{milk_price_time_units}
\alias{milk_price_time_units}
\title{Milk Price Time Units}
\format{
A tibble
}
\source{
NA
}
\usage{
milk_price_time_units
}
\description{
Conversion factors for the price of
milk, can both be a time or a per
unit volume
}
\keyword{datasets}
