% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LivestockCalculations.R
\name{milk_price_time_units_conversion}
\alias{milk_price_time_units_conversion}
\title{Milk Price Time Units Conversion}
\usage{
milk_price_time_units_conversion(
  units_column,
  sold_amount_column,
  units,
  unit_conversions
)
}
\arguments{
\item{units_column}{The columns containing the original units}

\item{sold_amount_column}{The amounts of milk which was sold}

\item{units}{A vector of units for which you have conversions}

\item{unit_conversions}{Conversion factors for the units listed above}
}
\value{

}
\description{
RHoMIS milk price units can come in both times and per
litre units. This function goes
}
