% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LivestockCalculations.R
\name{milk_proportions_all}
\alias{milk_proportions_all}
\title{Milk Proportions All}
\usage{
milk_proportions_all(data)
}
\arguments{
\item{data}{Data containing livestock loops necessary to calculate milk sold and consumed proportions}
}
\value{

}
\description{
A function to calculate the numeric proportions of milk
sold and milk consumed
}
