% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LivestockCalculations.R
\name{milk_sold_and_consumed_calculations}
\alias{milk_sold_and_consumed_calculations}
\title{Milk Sold and Consumed}
\usage{
milk_sold_and_consumed_calculations(data)
}
\arguments{
\item{data}{RHoMIS data including livestock loop information}
}
\value{

}
\description{
Function to calculate the amounts of milk sold
and consumed in litres.
}
