% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LivestockCalculations.R
\name{milk_swap_per_animal_units}
\alias{milk_swap_per_animal_units}
\title{Milk Swap Per Animal Units}
\usage{
milk_swap_per_animal_units(units_column, number_of_animals_milked_column)
}
\arguments{
\item{units_column}{The original column containing milk yield units}

\item{number_of_animals_milked_column}{A column containing the number of animals
milked.}
}
\description{
Some of RHoMIS milk yields are based on yield per animal.
This function converts yield per animal into a total milk yield.
  
Rpackage file: LivestockCalculations.R
}
