% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CleanColumnNames.R
\name{modify_all_loop_column_names}
\alias{modify_all_loop_column_names}
\title{Modify all types of loop in RHoMIS}
\usage{
modify_all_loop_column_names(column_names, repeat_columns)
}
\arguments{
\item{column_names}{A list of all of the column names which}

\item{repeat_columns}{A list of all of the types of repeat column which need to be changed}
}
\value{

}
\description{
The RHoMIS survey is structured using a series of important loops. These are often labelled "crop_repeat[x]".
It is important to be modify all of these loops to have a simpler column name format for later analysis
}
