% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CleanColumnNames.R
\name{modify_loop_name}
\alias{modify_loop_name}
\title{Modify Loop Name}
\usage{
modify_loop_name(column_name, loop_type)
}
\arguments{
\item{loop_type}{}
}
\value{

}
\description{
Many of the variables in RHoMIS are collected through a looping structure
}
