% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{modules}
\alias{modules}
\title{Modules}
\format{
A tibble
}
\source{
NA
}
\usage{
modules
}
\description{
A dataset describing different survey modules, and the functions which can be used to process them
}
\keyword{datasets}
