% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetaData.R
\name{new_indicator}
\alias{new_indicator}
\title{New Indicator}
\usage{
new_indicator(
  indicator_name,
  output_format = c("column", "loop", "table"),
  description,
  individual_columns_required = list(),
  loop_columns_required = list(),
  conversion_tables_required = list(),
  api_data_required = list(),
  indicators_required = list(),
  function_calculated,
  search_term
)
}
\arguments{
\item{indicator_name}{The name of the new indicator}

\item{output_format}{The shape of the indicator (column, loop or table)}

\item{description}{A description of the indicator}

\item{individual_columns_required}{Which individual columns are required (directly) for the RHoMIS dataset}

\item{loop_columns_required}{Which looped columns are required}

\item{conversion_tables_required}{Which conversion tables are required}

\item{api_data_required}{Which api connection data is required}

\item{indicators_required}{Other indicators which are required to calculate this one}

\item{function_calculated}{The function where this indicator is calculated}

\item{search_term}{The search term you can use to find where this indicator is calculated}
}
\description{
Create a new indicator
}
