% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NTFP.R
\name{ntfp_calories_and_values}
\alias{ntfp_calories_and_values}
\title{NTFP Calories and Values}
\usage{
ntfp_calories_and_values(tree_aid_df, price_conversions, calorie_conversions)
}
\arguments{
\item{calorie_conversions}{}
}
\description{
NTFP Calories and Values
}
