% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NTFP.R
\name{ntfp_total_individual}
\alias{ntfp_total_individual}
\title{NTFP Totals}
\usage{
ntfp_total_individual(
  tree_aid_df,
  fp_products = fp_products,
  income = F,
  value = F,
  calories = F,
  processed_income = F,
  processed_value = F,
  processed_calories = F,
  warning_message
)
}
\arguments{
\item{tree_aid_df}{A tree aid dataset}

\item{fp_products}{Forest products}

\item{income}{Whether to calculate income only}

\item{value}{Whether to calculate value only}

\item{calories}{Whether to calculate calories only}

\item{processed_income}{Whether to calculate processed income only}

\item{processed_value}{Whether to calculate processed value consumed only}

\item{processed_calories}{Whether to calculate processed calories consumed only}

\item{warning_message}{Warning message to give for any missing columns}
}
\description{
Calculate total ntfp income, value consumed,
calories consumed, processed income, value
of processed ntfp consumed, processed
calories consumed
}
