% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NTFP.R
\name{ntfp_totals}
\alias{ntfp_totals}
\title{NTFP Totals}
\usage{
ntfp_totals(tree_aid_df, indicator_df, fp_products = fp_products)
}
\arguments{
\item{fp_products}{}
}
\description{
Calculate total ntfp income, value consumed,
calories consumed, processed income, value
of processed ntfp consumed, processed
calories consumed
}
