% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetaData.R
\name{plot_dependency_network}
\alias{plot_dependency_network}
\title{Plot Dependency Network}
\usage{
plot_dependency_network(indicator_name, indicator_list, type = "horizontal")
}
\arguments{
\item{indicator_list}{}
}
\description{
Plot the dependency network of RHoMIS Indicators
}
