% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetaData.R
\name{plot_function_dependency}
\alias{plot_function_dependency}
\title{Add function to List}
\usage{
plot_function_dependency(
  function_list = function_list,
  function_name,
  type = "horizontal"
)
}
\arguments{
\item{function_list}{The list of functions}

\item{function_name}{The name of the function}

\item{type}{How to plot network}
}
\description{
Add function to List
}
