% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ppi_limits}
\alias{ppi_limits}
\title{PPI limits}
\format{
A tibble
}
\source{
A sample rhomis dataset
}
\usage{
ppi_limits
}
\description{
A dataset indicating which PPI thresholds
were used by each country
}
\keyword{datasets}
