% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ppi_score_card}
\alias{ppi_score_card}
\title{PPI Scorecard}
\format{
A tibble
}
\source{
A sample rhomis dataset
}
\usage{
ppi_score_card
}
\description{
A dataset used to convert from PPI score
to PPI likelihood for each individual country
}
\keyword{datasets}
