% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LivestockCalculations.R
\name{price_per_livestock}
\alias{price_per_livestock}
\title{Calculate Livestock Price}
\usage{
price_per_livestock(data)
}
\arguments{
\item{data}{Dataset containing livestock sold and livestock sale income}
}
\description{
Calculate the prices of whole livestock sold
}
\details{
Rpackage file: LivestockCalculations.R
}
