% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenerateMockResponses.R
\name{print_chunk_with_tab_spacing}
\alias{print_chunk_with_tab_spacing}
\title{Print Chunk with Tab Spacing}
\usage{
print_chunk_with_tab_spacing(xml_string, xml_level)
}
\arguments{
\item{xml_string}{The string which needs extra tabs
at the beginning}

\item{xml_level}{The number of tabs to be added at
the beginning of all lines}
}
\description{
When a xml chunk is created, say for a looped
set of questions, we can add the required
number of tabs at the beginning of each line
}
\details{
Rpackage file: GenerateMockResponses.R
}
