% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunAll.R
\name{processCentralProject}
\alias{processCentralProject}
\title{Process central Project and save}
\usage{
processCentralProject(
  central_url,
  central_email,
  central_password,
  project_name,
  form_name,
  form_version,
  database
)
}
\arguments{
\item{central_url}{The URL of the ODK central server}

\item{central_email}{The email of for your ODK central account}

\item{central_password}{The password for the ODK central account}

\item{project_name}{The name of the project you would like to process}

\item{form_name}{The name of the form you would like to process}

\item{form_version}{The verion of the form you would like to process}

\item{database}{The name of the database you would like to save your data to}
}
\value{

}
\description{
Process a project from ODK central and save it
to a mongoDB database
}
