% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunAll.R
\name{processData}
\alias{processData}
\title{Process RHoMIS data}
\usage{
processData(
  extractUnitsOnly = T,
  calculateInitialIndicatorsOnly = F,
  calculateFinalIndicatorsOnly = F,
  dataSource = c("csv", "central"),
  outputType = c("csv", "mongodb"),
  base_path = "./",
  dataFilePath = NULL,
  id_type = c("string", "column"),
  proj_id,
  form_id,
  overwrite = F,
  central_url = NULL,
  central_email = NULL,
  central_password = NULL,
  project_name = NULL,
  form_name = NULL,
  form_version = NULL,
  database = NULL,
  draft = NULL,
  central_test_case = FALSE,
  repeat_columns = pkg.env$repeat_columns,
  gender_categories = pkg.env$gender_categories
)
}
\arguments{
\item{extractUnitsOnly}{Whether or not to only extract units (TRUE/FALSE)}

\item{calculateInitialIndicatorsOnly}{Whether or not to only calcu,ate initial indicators (TRUE/FALSE)}

\item{calculateFinalIndicatorsOnly}{Whether or not to only calculate final indicators}

\item{dataSource}{The type of RHoMIS data being fed into the
calculations, whether a local csv file or data from ODK central.
Options "csv" or "central".}

\item{outputType}{The type of output to produce (options are "csv" or "mongodb")}

\item{dataFilePath}{The file to the data (csv format).
ONLY RELEVANT IF "dataSource" WAS "local".}

\item{id_type}{The type of ID you would like to use ("string" or "column")}

\item{proj_id}{An ID for your project}

\item{form_id}{An ID for your form}

\item{overwrite}{Whether or not to overwrite the converted values}

\item{central_url}{The url of the ODK-central server you are using.
ONLY RELEVANT IF "dataSource" WAS "central".}

\item{central_email}{The email of the ODK-central account you are using.
ONLY RELEVANT IF "dataSource" WAS "central"}

\item{central_password}{The password of the ODK-central account you are using.
ONLY RELEVANT IF "dataSource" WAS "central".}

\item{project_name}{The name of the ODK-central project you are processing.
ONLY RELEVANT IF "dataSource" WAS "central".}

\item{form_name}{The name of the ODK-central form you are processing.
ONLY RELEVANT IF "dataSource" WAS "central".}

\item{form_version}{The version of the ODK-central form you are processing.
ONLY RELEVANT IF "dataSource" WAS "central".}

\item{database}{The name of the database you would like to save results to}

\item{draft}{Whether or not the ODK for you war working with is a draft
or a final version. Only relevant if you are processing a project from ODK central}

\item{central_test_case}{This flag is used for running a test-sample dataset from ODK the inst/sample_central_project/ folder}

\item{repeat_columns}{The columns which are looped in the datasets being processed}
}
\value{

}
\description{
A main function that can be used to process rhomis
data. Whether the dataset comes from a local csv
or from ODK central.
}
\details{
RHoMIS datasets go through 4 stages of processing:

1. Initial Cleaning and Extracting Units
2. Calculation of initial indicators
3. Calculation of final indicators, including food availability, gender, and value of products consumed
}
\examples{
If local csv specified, load the csv
add some identification columns and
clean the column names
If central dataset specified,
identify the relevant projects,
and load the zip files.
add some identification columns and
clean the column names
There are some extra central columns
which are problematic, these need to be removed
Make an empty indicator dataset with
matching ID columns
Run all of the preliminary calculations that can
be done without price verification and without
verification of calory values
This function can be found in the
R/redirectModules.R file.
From this function we receive a list of
data frames. These include processed_data,
indicator_data, and extra_outputs
If the desired output format is csv,
Write the processed data, indicator
data and extra outputs
will be written to the relevant files


NEED TO SIMPLIFY
Below we write all of the data
which is processed during the initial calculations
}
