% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReformatLoops.R
\name{prop_or_na}
\alias{prop_or_na}
\title{Proportion or NA}
\usage{
prop_or_na(item)
}
\arguments{
\item{item}{A list of the people controlling a particular resource}
}
\value{

}
\description{
For a particular gender control item, e.g. "female_adult male_adult" this function
works out whether to return NA or a proportion. In this case the proportion controlled
by each person would be 0.5
}
\examples{
item <- c("male_adult", "female_adult")
prop_or_na(item) #should return 0.5
}
