% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReformatLoops.R
\name{proportion_control_per_person}
\alias{proportion_control_per_person}
\title{Number Controlling Resource}
\usage{
proportion_control_per_person(item)
}
\arguments{
\item{item}{The specific string, for which we want to count the
number controlling the resource}
}
\value{
A proportion (0-1) for the people controlling resource
}
\description{
For the RHoMIS gender columns, we ask who in the household collects each resource.
This function is designed to be applied to each item (spicific row for specific column)
}
\examples{

column <- c("male_adult female_adult","male_adult",NA,"female_youth")
proportion_control_per_person(column)

}
