% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{proportion_conversions}
\alias{proportion_conversions}
\title{Proportion Conversions}
\format{
A tibble
}
\source{
NA
}
\usage{
proportion_conversions
}
\description{
Conversion factors for the string proportions
used to calculate the amounts of products sold
and consumed
}
\keyword{datasets}
