% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CurrencyConversion.R
\name{querywb_stats}
\alias{querywb_stats}
\title{Query World Bank Statistics}
\usage{
querywb_stats(year, country_code)
}
\arguments{
\item{year}{The year you are hoping to query}

\item{country_code}{The two letter ISO country code}
}
\value{

}
\description{
Find out PPP conversions based on the
world bank statistics
}
