% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redirectModules.R
\name{read_folder_of_csvs}
\alias{read_folder_of_csvs}
\title{Read Folder of CSVs}
\usage{
read_folder_of_csvs(folder = "./")
}
\arguments{
\item{folder}{The folder containing the tables to load}
}
\description{
A function
 
Rpackage file: redirectModules.R
}
