% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MongoDB.R
\name{read_in_db_dataset}
\alias{read_in_db_dataset}
\title{Read in DB Dataset}
\usage{
read_in_db_dataset(
  collection = "data",
  database,
  project_name,
  form_name,
  data_set_name
)
}
\arguments{
\item{collection}{The collection containing the data needed}

\item{database}{The name of the database}

\item{project_name}{The name of the project which you are searching for}

\item{form_name}{The name of the form you are searching for}

\item{data_set_name}{The name of the dataset.}
}
\description{
A function for reading and unwinding
data in the RHoMIS database.
  
Rpackage file: MongoDB.R
}
