% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConnectToCentral.R
\name{remove_extra_central_columns}
\alias{remove_extra_central_columns}
\title{Remove Extra Central Columns}
\usage{
remove_extra_central_columns(data)
}
\arguments{
\item{data}{ODK central data with the excess columns}
}
\value{

}
\description{
Data downloaded from ODK central has extra columns
with no additional information. These extra columns
follow a particular pattern. This function removes these
columns based on this pattern
}
