% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CleanRawData.R
\name{replace_crop_and_livestock_other}
\alias{replace_crop_and_livestock_other}
\title{Replace crop and livestock other}
\usage{
replace_crop_and_livestock_other(data)
}
\arguments{
\item{data}{A whole rhomis dataset}
}
\value{

}
\description{
Replace "other1", "other2", and "other3" text
entries in the crop and livestock names
}
