% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunAll.R
\name{replace_infinite}
\alias{replace_infinite}
\title{Replace Infinite}
\usage{
replace_infinite(column)
}
\arguments{
\item{column}{The column where infinite values need to be replaced}
}
\description{
Replace infinite values with NA in a specific column
 
Rpackage file: RunAll.R
}
