% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CleanRawData.R
\name{replace_name_column_with_other}
\alias{replace_name_column_with_other}
\title{Replace Name Column with Other}
\usage{
replace_name_column_with_other(
  main_column,
  other_column1,
  other_column2,
  other_column3
)
}
\arguments{
\item{main_column}{The column you are searching for "other" values in (e.g. crop_name)}

\item{other_column1}{The first "other" column you are searching through, e.g. "crops_other1"}

\item{other_column2}{The second "other" column you are searching through, e.g. "crops_other2"}

\item{other_column3}{The third "other" column you are searching through, e.g. "crops_other3"}
}
\value{

}
\description{
Replace Name Column with Other
}
