% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CleanRawData.R
\name{replace_unit_column_with_other_single}
\alias{replace_unit_column_with_other_single}
\title{Replace unit with Single Value}
\usage{
replace_unit_column_with_other_single(unit_column, unit_other_column)
}
\arguments{
\item{unit_column}{The main unit column}

\item{unit_other_column}{The column containing the free text entry}
}
\description{
For two columns, identify where
"other" is written, and replace
it with th appropriate text entry
}
\details{
Rpackage file: CleanRawData.R
}
