% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CleanRawData.R
\name{replace_unit_list}
\alias{replace_unit_list}
\title{Replace unit list}
\usage{
replace_unit_list(list_to_convert, unit_conv_tibble)
}
\arguments{
\item{list_to_convert}{The list of character units you need to convert}

\item{unit_conv_tibble}{The tibble of unit conversions}
}
\value{
A list of numerically converted units
}
\description{
Replacing all of the character units in an R list
}
\examples{
list_to_convert <- c("kg", "sacks_50kg","other_unit",NA, NA, "wheel_barrows_100kg","litres")
unit_conv_tibble <- tibble::as_tibble(list(unit=c("kg", "sacks_50kg", "wheel_barrows_100kg","litres"),
                                   conversion_factors= c(1,50,100,1)))
replace_unit_list(list_to_convert,unit_conv_tibble)
}
