% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CleanRawData.R
\name{replace_unit_with_conversion_factor}
\alias{replace_unit_with_conversion_factor}
\title{Replace unit with conversion factor}
\usage{
replace_unit_with_conversion_factor(item_to_convert, unit_conv_tibble)
}
\arguments{
\item{item_to_convert}{The character unit which needs to be converted}

\item{unit_conv_tibble}{A table of units, and their numerical conversion factor. With the column names "unit" and "conversion_factors"}
}
\value{
A numeric conversion of the original character unit
}
\description{
A function to replace individual units with their appropriate conversion factors
}
\examples{

item_to_convert <- "wheel_barrows_100kg"
unit_conv_tibble <- as_tibble(list(unit=c("kg", "sacks_50kg", "wheel_barrows_100kg","litres"), conversion_factors= c(1,50,100,1)))
replace_unit_with_conversion_factor(item_to_convert, unit_conv_tibble)
}
