% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CleanRawData.R
\name{replace_units_with_other_all}
\alias{replace_units_with_other_all}
\title{Replacing Units with other Units}
\usage{
replace_units_with_other_all(data)
}
\arguments{
\item{data}{The whole rhomis dataset}
}
\value{

}
\description{
In the rhomis survey text entries are recorded
as "other". The actual text is stored in another variable.
This can be problematic for analysis. This function brings
the "other" units into where they are needed.
}
