% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhomisApiConnection.R
\name{rhomis_authenticate}
\alias{rhomis_authenticate}
\title{RHoMIS Authenticate}
\usage{
rhomis_authenticate(base_url, email, password)
}
\arguments{
\item{base_url}{The URL for the api you are interacting with}

\item{email}{The email address of your account}

\item{password}{The password for your account}
}
\value{

}
\description{
Authenticate user for interacting with the
RHoMIS data api and authentication AP
}
