% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redirectModules.R
\name{run_preliminary_calculations}
\alias{run_preliminary_calculations}
\title{Run preliminary calculations}
\usage{
run_preliminary_calculations(
  rhomis_data,
  gender_categories = pkg.env$gender_categories,
  units
)
}
\arguments{
\item{rhomis_data}{A tibble of rhomis_data}

\item{gender_categories}{The gender categories present in the data which is to be processed}

\item{units}{A nested list containing all of the unit conversion tibbles}
}
\description{
Rpackage file: redirectModules.R
}
