% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sample_raw_data_set}
\alias{sample_raw_data_set}
\title{Sample Raw Data Set}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 10 rows and 797 columns.
}
\source{
NA
}
\usage{
sample_raw_data_set
}
\description{
A sample raw rhomis dataset
}
\keyword{datasets}
