% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetaData.R
\name{save_conversion_metadata}
\alias{save_conversion_metadata}
\title{A set of functions designed to help save
and store metadata in an organised way}
\usage{
save_conversion_metadata(unit_name, projectID, formID)
}
\description{
A set of functions designed to help save
and store metadata in an organised way
}
