% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-mongodb.R
\name{save_data_set_to_db}
\alias{save_data_set_to_db}
\title{Save Dataset to DB}
\usage{
save_data_set_to_db(
  data,
  data_type,
  database = "rhomis",
  url = "mongodb://localhost",
  projectID,
  formID
)
}
\arguments{
\item{data}{The data to save}

\item{data_type}{The type of data to save (e.g. cropData, indicatorData...)}

\item{database}{The name of the database to save it to}

\item{url}{The url of the database}

\item{projectID}{The name of the project containing the data}

\item{formID}{The ID of the form}
}
\description{
Save a dataset to the MongoDB database
  
Rpackage file: utils-mongodb.R
}
