% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UnitsAndConversions.R
\name{save_initial_units}
\alias{save_initial_units}
\title{Save Initial Units}
\usage{
save_initial_units(
  database = "rhomis",
  url = "mongodb://localhost",
  projectID = "core_units",
  formID = "core_units",
  units_and_conversions = list(crop_price_units = crop_price_units, crop_yield_units =
    crop_yield_units, eggs_amount_units = eggs_amount_units, eggs_price_time_units =
    eggs_price_time_units, honey_amount_units = honey_amount_units, land_area_units =
    land_area_units, livestock_weights = livestock_weights, milk_amount_units =
    milk_amount_units, milk_price_time_units = milk_price_time_units, ppi_limits =
    ppi_limits, ppi_score_card = ppi_score_card, proportion_conversions =
    proportion_conversions)
)
}
\arguments{
\item{database}{The database where units are to be saved}

\item{url}{The URL of the database where units are to be saved, the default mongoDB url is set as default argument.}

\item{projectID}{The ID of the project for saving the units}

\item{formID}{The ID of the form for saving project units}

\item{units_and_conversions}{A list of all of the unit and conversions (list of tibbles)}
}
\value{

}
\description{
Save Initial Units
}
