% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-mongodb.R
\name{save_list_of_df_to_db}
\alias{save_list_of_df_to_db}
\title{Save List of Data Frames to the RHoMIS Database}
\usage{
save_list_of_df_to_db(
  list_of_df,
  projectID,
  formID,
  database = "rhomis",
  url = "mongodb://localhost"
)
}
\arguments{
\item{list_of_df}{The list of dataframes}

\item{projectID}{The name of the project}

\item{formID}{The name of the form}

\item{database}{The name of the database4}

\item{url}{The url of the database}
}
\description{
Take a list of output datasets and save them all to the
RHoMIS mongoDB
}
\details{
Rpackage file: utils-mongodb.R
}
