% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UnitsAndConversions.R
\name{save_multiple_conversions}
\alias{save_multiple_conversions}
\title{Save Multiple conversions}
\usage{
save_multiple_conversions(
  database = "rhomis",
  url = "mongodb://localhost",
  projectID = "core_units",
  formID = "core_units",
  conversion_data,
  conversion_types,
  collection,
  converted_values = F
)
}
\arguments{
\item{database}{The database which will store the units}

\item{url}{The url of the database storing the units}

\item{projectID}{The name of the project the units are related to}

\item{formID}{The name of the form the projects are related to}

\item{conversion_data}{A list of tibbles of the conversion factors to be used}

\item{conversion_types}{Which types of conversion is being saves (e.g crop yield, PPI, HDDS etc...)}

\item{collection}{The mongodb collection to save the units}

\item{converted_values}{Whether saving values which user has already converted}
}
\value{

}
\description{
Save Multiple conversions
}
