% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UnitsAndConversions.R
\name{save_set_of_conversions}
\alias{save_set_of_conversions}
\title{Saving Set of Conversions}
\usage{
save_set_of_conversions(
  database = "rhomis",
  url = "mongodb://localhost",
  projectID = "core_unit",
  formID = "core_unit",
  conversions,
  conversion_type,
  collection,
  converted_values = F
)
}
\arguments{
\item{database}{The database which will store the units}

\item{url}{The url of the database storing the units}

\item{projectID}{The name of the project the units are related to}

\item{formID}{The name of the form the projects are related to}

\item{conversions}{A table of the conversion factors to be used}

\item{conversion_type}{Which type of conversion is being saves (e.g crop yield, PPI, HDDS etc...)}

\item{collection}{The mongodb collection to save the units}

\item{converted_values}{Whether saving values which user has already converted}
}
\value{

}
\description{
Save a single type of conversion factor into the conversion factor database
}
