% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenerateMockResponses.R
\name{select_multiple}
\alias{select_multiple}
\title{Select Multiple}
\usage{
select_multiple(list_to_sample)
}
\arguments{
\item{list_to_sample}{The list of objects which need to be
sampled}
}
\description{
A function for selecting multiple choices
from a list at random
}
