% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-mongodb.R
\name{set_project_tag_to_true}
\alias{set_project_tag_to_true}
\title{Set Project Tag to True}
\usage{
set_project_tag_to_true(
  database = "rhomis",
  url = "mongodb://localhost",
  projectID,
  formID,
  project_tag
)
}
\arguments{
\item{database}{Database to add it to}

\item{url}{URL of the database}

\item{projectID}{ID of the project you are adding}

\item{formID}{The id of the form being you are adding}

\item{project_tag}{Which project tag needs to be set to true}
}
\description{
RHoMIS data goes through several
processing stages. Where units are
extracted, then verified, then 
prices verified. This function allows
us to tick off these stages, letting
the system know what has been verified and what has not.
   
Rpackage file: utils-mongodb.R
}
