% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CleanColumnNames.R
\name{shorten_individual_column_name}
\alias{shorten_individual_column_name}
\title{This script includes a series of functions which can be
used to help clean column names of data collected with RHoMIS.
Shorten Individual Name}
\usage{
shorten_individual_column_name(column_name, seperator)
}
\arguments{
\item{column_name}{The specific item which needs shortening}

\item{seperator}{The character separating parts of the item}
}
\value{
Shortened string of the essential column name
}
\description{
Used to abbreviate a single column name
based on its last value. Typical RHoMIS
column names come in the form "xxxx/xxxxx/xxxxx/important_name".
This function helps to extract only the final value
"important name".
}
\examples{

long_name <- "xxxx/xxxxx/xxxxx/important_name"
shortened_name <- shorten_individual_column_name(
  column_name = long_name,
  seperator = "/"
)
# Shortened name will be equal "important_name"
}
