% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CleanColumnNames.R
\name{shorten_multiple_column_names}
\alias{shorten_multiple_column_names}
\title{Shorten Multiple Column Names}
\usage{
shorten_multiple_column_names(long_names, seperator)
}
\arguments{
\item{long_names}{The list of column names which need to be shortened}

\item{seperator}{The string seperating parts of each column name}
}
\value{
A list of shortened column names
}
\description{
Used to abbreviate multiple column names based on their last values. Typical RHoMIS column names come
in the for "xxxx/xxxxx/xxxxx/important_name". This function helps to extract only the final value
"important name". It will do this for multiple column names
}
\examples{

long_names <- c("xxxx/xxxxx/xxxxx/important_name_1",
                "xxxx/xxxxx/xxxxx/important_name_2",
                "xxxx/xxxxx/xxxxx/important_name_3")
seperator <- "/"

shortened_names <- shorten_multiple_column_names(long_names, seperator)
# shortened_names will equal c("important_name_1","important_name_2", "important_name_3")
}
