% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReformatLoops.R
\name{split_gender_columns}
\alias{split_gender_columns}
\title{Split Gender columns}
\usage{
split_gender_columns(column, gender_categories = pkg.env$gender_categories)
}
\arguments{
\item{column}{The column that needs to be split}

\item{gender_categories}{The gender categories to
be examined in the rhomis dataset}
}
\description{
A function to split gender columns into multiple
columns based on the different groups
}
\details{
Rpackage file: ReformatLoops.R
}
