% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataReformatting.R
\name{split_string_categories_to_dummy}
\alias{split_string_categories_to_dummy}
\title{Split string to dummy columns}
\usage{
split_string_categories_to_dummy(x, seperator)
}
\arguments{
\item{x}{A vector which needs to be split}

\item{seperator}{The separating character}
}
\description{
Many RHoMIS columns come in the format c("cropA cropB cropC", "cropD cropA").
This function helps split these into more interperatable dummy columns
}
\details{
Rpackage file: DataReformatting.R
}
