% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConnectToCentral.R
\name{submit_xml_data}
\alias{submit_xml_data}
\title{Submit XML Data}
\usage{
submit_xml_data(
  xml_string,
  central_url,
  central_email,
  central_password,
  projectID,
  formID,
  isDraft
)
}
\arguments{
\item{xml_string}{The XML string for the data submission you
would like to make}

\item{central_url}{The url of the ODK central server}

\item{central_email}{The email of your ODK central account}

\item{central_password}{The password to your ODK central account}

\item{projectID}{The ID of the project you are looking at.
To get a list of project, see the
"get_projects" function}

\item{formID}{The ID of the form containing the submissions you are looking at.
To get the list of forms see the "get_forms" function.}

\item{isDraft}{Whether the form you are submitting to is a draft or not}
}
\description{
Rpackage file: ConnectToCentral.R
}
