% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LivestockCalculations.R
\name{swap_livestock_heads_other}
\alias{swap_livestock_heads_other}
\title{Swap Livestock Heads With Other}
\usage{
swap_livestock_heads_other(data)
}
\arguments{
\item{data}{A rhomis dataset}
}
\description{
TLU column names include options for
"other", this is not interperatable or
easy to use for analysis. This function addresses
this
}
\details{
Rpackage file: LivestockCalculations.R
}
