% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataReformatting.R
\name{switch_column_names_and_add_categories_for_specific_project}
\alias{switch_column_names_and_add_categories_for_specific_project}
\title{Switch Column Names and Add Categories For a Project}
\usage{
switch_column_names_and_add_categories_for_specific_project(
  data,
  conversion_tibble,
  id_rhomis_dataset
)
}
\arguments{
\item{data}{Dataset with categories to be merged}

\item{conversion_tibble}{A tibble of conversion factors}

\item{id_rhomis_dataset}{The id of the rhomis dataset to
be converted}
}
\value{

}
\description{
A function for changing the column names
e.g, where the column names represent crops,
and merging categories.
}
