% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataReformatting.R
\name{switch_column_names_and_merge_categories}
\alias{switch_column_names_and_merge_categories}
\title{Switch Column Names and Add Categories For a Project}
\usage{
switch_column_names_and_merge_categories(
  data,
  conversion_tibble,
  by_project = T
)
}
\arguments{
\item{data}{Dataset with categories to be merged}

\item{conversion_tibble}{A tibble of conversion factors
be converted}

\item{by_project}{Should}
}
\value{

}
\description{
A function for changing the column names
e.g, where the column names represent crops,
and merging categories.
}
