% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CleanRawData.R
\name{switch_units}
\alias{switch_units}
\title{Switch Units}
\usage{
switch_units(data_to_convert, units, conversion_factors)
}
\arguments{
\item{data_to_convert}{This list or data frame which needs to be converted}

\item{units}{A list of the character units which need to be converted (see example)}

\item{conversion_factors}{A list of converstion factors for the unit list (see example)}
}
\value{
Returns the data (in its original format) with the units in their numerical format
}
\description{
In RHoMIS, sometimes lists or data frames contain units which need to be converted
from a character format (e.g. "kg","sacks_100kg") into a numeric conversion (e.g. 1, 100).
This function does just that.
}
\examples{
# Converting a list
units <- c("kg", "sacks_50kg", "wheel_barrows_100kg","litres")
conversion_factors <- c(1,50,100,1)
list_to_convert <- c("kg", "sacks_50kg","other_unit",NA, NA, "wheel_barrows_100kg","litres")
switch_units(list_to_convert, units, conversion_factors)
# Converting a data frame or tibble
units <- c("kg", "sacks_50kg", "wheel_barrows_100kg","litres")
conversion_factors <- c(1,50,100,1)
tibble_to_convert <- as_tibble(list("maize"=c("kg", "other_random_unit","wheel_barrows_100kg"),
                                    "cassava"=c("sacks_50kg",NA,"another_random_unit"),
                                    "banana"=c("bunches", "wheel_barrows_100kg",NA)))
switch_units(tibble_to_convert, units, conversion_factors)

}
