% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TotalIncomes.R
\name{total_and_off_farm_incomes}
\alias{total_and_off_farm_incomes}
\title{Calculating off-farm income}
\usage{
total_and_off_farm_incomes(data, total_crop_income, total_livestock_income)
}
\arguments{
\item{data}{RHoMIS data, including column of offfarm income}

\item{total_crop_income}{Total income from crops list}

\item{total_livestock_income}{Total income from livestock list
proportions}
}
\description{
Calculating off-farm income based on total income proportions
}
\details{
Rpackage file: TotalIncomes.R
}
