% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TotalIncomes.R
\name{total_crop_income}
\alias{total_crop_income}
\title{Total Crop Income}
\usage{
total_crop_income(data)
}
\arguments{
\item{data}{RHoMIS data that has been processed using
the "crop_calculations_all" function}
}
\value{

}
\description{
Calculate the total crop income from processed
RHoMIS data
}
