% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TotalIncomes.R
\name{total_livestock_income}
\alias{total_livestock_income}
\title{Total Livestock Income}
\usage{
total_livestock_income(data)
}
\arguments{
\item{data}{RHoMIS data that has been processed using
the "livestock_calculations_all" function}
}
\value{

}
\description{
Calculate total livestock income from
processed RHoMIS data
}
