% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redirectModules.R
\name{value_gender_fa_calculations}
\alias{value_gender_fa_calculations}
\title{Value, Gender, and Food Availability Calculations}
\usage{
value_gender_fa_calculations(
  processed_data,
  indicator_data,
  calorie_conversions,
  prices
)
}
\arguments{
\item{processed_data}{RHoMIS Processed Dataset}

\item{prices}{A list of tibbles on mean prices}

\item{crop_data}{A list of tibbles on crop data}

\item{livestock_data}{A list of tibbles on livestock data}

\item{off_farm_data}{Information on prices}
}
\value{

}
\description{
Value, Gender, and Food Availability Calculations
}
