% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ValueCalculations.R
\name{value_or_calorie_calculations_item_consumed}
\alias{value_or_calorie_calculations_item_consumed}
\title{Value or Calorie Calculations}
\usage{
value_or_calorie_calculations_item_consumed(
  data,
  name_column,
  amount_consumed_column,
  conversion_tibble,
  price_column_name,
  converted_column_name
)
}
\arguments{
\item{data}{The whole processed rhomis dataset}

\item{name_column}{The loop column containing the name pattern (e.g "crop_name_1", "crop_name_2", "crop_name_3" would be "crop_name")}

\item{amount_consumed_column}{The loop column containing the consumed amount pattern (e.g. "crop_consumed_kg_per_year_1"... would be "crop_consumed_kg_per_year".)}

\item{conversion_tibble}{The dataframe for the conversion (e.g. price or calorie conversion sheet)}

\item{price_column_name}{The name you would like to include when adding the mean prices to the dataframe (e.g. mean_crop_price_per_kg)}

\item{converted_column_name}{The name of the converted column you would like to give (e.g.value_crop_consumed_lcu)}
}
\value{

}
\description{
Calculate the value of an item in terms of calories or monetary value.
}
