% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenderControlAll.R
\name{widen_gender_columns}
\alias{widen_gender_columns}
\title{Widen Gender Columns}
\usage{
widen_gender_columns(processed_data, gender_group, column_pattern)
}
\arguments{
\item{processed_data}{A processed dataset, containing the gender columns}

\item{gender_group}{Which gender group you're interested in}

\item{column_pattern}{The pattern of the column you are processing (e.g. for "male_youth_crop_consumed_kg",
the pattern would be "crop_consumed_kg"}
}
\description{
Take gender categories and a processed dataset,
reformat the gender control variables into a more
useable format
}
\details{
Rpackage file: GenderControlAll.R
}
