% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConnectToCentral.R
\name{widen_individual_result}
\alias{widen_individual_result}
\title{Widen Individual Central Results}
\usage{
widen_individual_result(individual_central_item, column_headers)
}
\arguments{
\item{individual_central_item}{The individual list item subsetted
from a central query}

\item{column_headers}{The desired headers for the new table}
}
\value{

}
\description{
Widening a single central result into a simple tibble
}
\examples{
central_results <- list(
  list("id" = 1, "name" = "name1", "email" = "email1"),
  list("id" = 2, "name" = "name2", "email" = "email2"),
  list("id" = 3, "name" = "name3", "email" = "email3")
)
column_headers <- unique(names(unlist(central_results)))
individual_central_result <- central_results[2]
widen_individual_result(individual_central_result, column_headers)
}
