% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExtractNewValues.R
\name{write_core_values_to_convert_to_file}
\alias{write_core_values_to_convert_to_file}
\title{Write Core Values to File}
\usage{
write_core_values_to_convert_to_file(data, folder)
}
\arguments{
\item{data}{The data for which you want to
extract all of the new files.}

\item{folder}{The folder where you want to
store all of the values you need to convert.
Make sure the folder is correct and does not end in "/"}
}
\description{
This function extracts all of the new units and
names in a RHoMIS survey and writes them to a folder
You can select the folder that you would like
to write all of the files to.
}
\details{
Rpackage file: ExtractNewValues.R
}
\examples{
data <- tibble::tibble(
  crop_name_1 = c("maize", "cassava"),
  crop_name_2 = c(NA, "cucumber"),
  random_crop_name_2 = c("blue", "green"),
  livestock_name_1 = c("cow", "chicken"),
  livestock_name_2 = c("pig", NA),
  crop_yield_units_1 = c("kg", "sacks"),
  crop_yield_units_2 = c("wheelbarrow", NA),
  crop_yield_units_other_1 = c("other1", NA),
  crop_yield_units_other_2 = c("other2", "other3"),
  crop_sold_price_quantityunits_1 = c(NA, "price1"),
  crop_sold_price_quantityunits_2 = c("price2", "price3"),
  crop_price_quantityunits_other_1 = c(NA, "crop_price_1"),
  crop_price_quantityunits_other_2 = c(NA, "crop_price_2"),
  crop_price_quantityunits_other_3 = c(NA, "crop_price_3"),
  crop_price_quantityunits_other_4 = c(NA, NA),
  unitland = c("acre", NA),
  areaunits_other = c("area1", "area2"),
  unitland_owned = c("unitland1", "unitland2"),
  unitland_rentin = c("renty1", "renty2"),
  unitland_rentout = c("rent1", "rent5"),
  milk_units_1 = c("milk1", NA),
  milk_units_2 = c("milk4", "milk5"),
  milk_amount_units_other_1 = c("milkoth1", "milkoth2"),
  milk_amount_units_other_2 = c(NA, "milkoth3"),
  milk_sold_price_timeunits_1 = c("mspt1", "mspt3"),
  milk_sold_price_timeunits_2 = c(NA, "mspt5"),
  milk_amount_time_units_other_1 = c("mspto1", NA),
  milk_amount_time_units_other_2 = c("mspto2", "mspto3"),
  bees_honey_production_units_1 = c("hnyprod1", "hnyprod2"),
  bees_honey_production_units_2 = c(NA, "hnyprod3"),
  bees_honey_production_units_other_1 = c("hnyprodoth1", NA),
  bees_honey_production_units_other_2 = c("hnyprodoth2", "hnyprodoth3"),
  eggs_units_1 = c("egg1", NA),
  eggs_units_2 = c("egg2", "egg3"),
  eggs_amount_units_other_1 = c("eggoth1", NA),
  eggs_amount_units_other_2 = c("eggoth2", "eggoth3"),
  eggs_sold_price_timeunits_1 = c("eggtime1", "eggtime2"),
  eggs_sold_price_timeunits_2 = c("eggtime3", NA),
  eggs_sold_price_timeunits_3 = c(NA, NA),
  eggs_sold_price_timeunits_other_1 = c("eggtimeoth1", NA),
  eggs_sold_price_timeunits_other_2 = c("eggtimeoth2", "eggtimeoth3"),
  fertiliser_units = c("fert1", "fert2"),
  fertiliser_units_other = c("fertoth1", "fertoth2")
)

# write_core_values_to_convert_to_file(data, "My/Folder)
}
