% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-create-downloadable-folder.R
\name{write_docs}
\alias{write_docs}
\title{Write Docs}
\usage{
write_docs(project_name, form_name)
}
\arguments{
\item{project_name}{The name of the project}

\item{form_name}{The name of the form}
}
\description{
Write a readme for a RHoMIS
dataset which has been automatically
generated
}
