% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-mongodb.R
\name{write_new_collection}
\alias{write_new_collection}
\title{Write a New Collection}
\usage{
write_new_collection(
  data_to_write,
  collection,
  database = "rhomis",
  url = "mongodb://localhost"
)
}
\arguments{
\item{data_to_write}{The dataframe that you want to write to file}

\item{collection}{The collection you want to collect the information from}

\item{database}{The database you are querying}

\item{url}{The url of the database you are querying}
}
\description{
Write a new table to a RHoMIS db database
  
Rpackage file: utils-mongodb.R
}
