% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExtractNewValues.R
\name{write_new_values_to_file}
\alias{write_new_values_to_file}
\title{Write units tibble to csv}
\usage{
write_new_values_to_file(specific_value, list_of_tibbles, folder)
}
\arguments{
\item{specific_value}{Which attribute are you
hoping to write to csv (e.g. "crop_yield_units")}

\item{list_of_tibbles}{What are all of the units
which you have extracted from the RHoMIS project.
See function "extract_new_core_units"}

\item{folder}{What is the folder you hope to write this to.
Do not include "/" at the end of the file path}
}
\value{
No return
}
\description{
Write the specific tibble to a csv.
}
\examples{

# Please make sure the correct folder exists for you
specific_value <- "egg_units"
new_values <- list(egg_units = c("egg1","egg2","egg3" ),
                   other_units = c("other1","other2","other3" ))
list_of_tibbles <- sapply(new_values, function(x) convert_new_values_to_tibble(x),simplify = FALSE)
folder <- "data/test_outputs/write_new_values_to_file"


#write_new_values_to_file(specific_value,list_of_tibbles,folder)

}
