% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UnitsAndConversions.R
\name{write_units_to_folder}
\alias{write_units_to_folder}
\title{Write Units to Folder}
\usage{
write_units_to_folder(list_of_df, folder = "./unit_conversions")
}
\arguments{
\item{list_of_df}{A list of dataframes, containing all
of the units and conversions}
}
\value{

}
\description{
Write all of the units and new names to a
folder where they can be checked and converted locally
}
