% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MongoDB.R
\name{add_data_to_project_list}
\alias{add_data_to_project_list}
\title{Add Data to Project List}
\usage{
add_data_to_project_list(
  data,
  collection,
  database = "rhomis",
  url = "mongodb://localhost"
)
}
\arguments{
\item{data}{Data to add}

\item{collection}{Collection to add it to}

\item{database}{Database to add it to}

\item{url}{URL of the database}
}
\value{

}
\description{
A method for submitting another project
to a list of projects.
}
