% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HouseholdSize.R
\name{calculate_MAE}
\alias{calculate_MAE}
\title{Calculate MAE}
\usage{
calculate_MAE(data)
}
\arguments{
\item{data}{The data containing all rhomis information}
}
\description{
A function to calculate the male adult equivalent of household size
}
\details{
Rpackage file: HouseholdSize.R
}
