% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunAll.R, R/redirectModules.R
\name{calculate_prices_and_indicator_local}
\alias{calculate_prices_and_indicator_local}
\title{Calculate prices and indicator local}
\usage{
calculate_prices_and_indicator_local(
  data,
  base_path = "./",
  units_path = "./converted_units/"
)

calculate_prices_and_indicator_local(
  data,
  base_path = "./",
  units_path = "./converted_units/"
)
}
\arguments{
\item{data}{A rhomis dataset, loaded with the function ()}

\item{base_path}{Path to the project folder}

\item{units_path}{Path to the folder of converted units}

\item{path_to_units}{}
}
\value{



}
\description{
Calculate prices and initial indicators locally
}
