% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redirectModules.R
\name{calculate_prices_and_indicator_local}
\alias{calculate_prices_and_indicator_local}
\title{Calculate prices and indicator Local}
\usage{
calculate_prices_and_indicator_local(
  data,
  base_path = "./",
  units_path = "converted_units/",
  gender_categories = pkg.env$gender_categories,
  units
)
}
\arguments{
\item{data}{The dataset which is to be processed}

\item{base_path}{Path to the project folder}

\item{units_path}{Path to the folder of converted units}

\item{gender_categories}{The gender categories used in this survey}

\item{units}{A nested list containing all of the unit conversion tibbles}
}
\description{
Calculate prices and initial indicators locally
}
\details{
Rpackage file: redirectModules.R
}
