% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-calculate-prices.R
\name{calculate_prices_csv}
\alias{calculate_prices_csv}
\title{Calculate Prices CSV}
\usage{
calculate_prices_csv(
  base_path = "./",
  file_path,
  id_type = c("string", "column"),
  proj_id,
  form_id,
  unique_id_col = "_uuid"
)
}
\arguments{
\item{base_path}{The folder where you want to save any outputs, usually your current working directory ("./")}

\item{file_path}{Filepath to a RHoMIS Survey CSV}

\item{id_type}{RHoMIS surveys have form and project IDs. Sometimes the form and project IDs are included as a column in the dataset (id_type="column"), or the IDs are specified by the user at the point of processing (id_type="string")}

\item{proj_id}{If ID type was string, this should be a string, if ID type was column, this should be a column name containing project IDs}

\item{form_id}{If ID type was string, this should be a string, if ID type was column, this should be a column name containing form IDs}

\item{unique_id_col}{The column containing unique household ids

Rpackage file: 02-calculate-prices.R}
}
\description{
Calculate Prices CSV
}
