% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalorieCalculations.R
\name{calorie_calculations}
\alias{calorie_calculations}
\title{Value Calculations}
\usage{
calorie_calculations(processed_data, indicator_data, calorie_conversions)
}
\arguments{
\item{mean_prices}{}
}
\value{

}
\description{
Value Calculations
}
