% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataContentsCheck.R
\name{check_columns_in_data}
\alias{check_columns_in_data}
\title{Check columns in data}
\usage{
check_columns_in_data(
  data,
  loop_columns = NULL,
  individual_columns = NULL,
  warning_message = NULL
)
}
\arguments{
\item{data}{The dataset which needs checking}

\item{loop_columns}{The columns which are contained in a loop}

\item{individual_columns}{The columns which appear individually}

\item{warning_message}{The warning message (perhaps the name of the function you are calling)
to indicate where the problem may be happening. A warning message will already be included
if the data is missing. This warning information can be supplied to give some context}
}
\value{

}
\description{
For many of the functions to work, we need to ensure that
the correct columns are present in the tibble. This
function checks whether the columns are present, both
for looped, and individual columns
}
