% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CleanColumnNames.R
\name{clean_column_names}
\alias{clean_column_names}
\title{Clean Column Names}
\usage{
clean_column_names(column_names)
}
\arguments{
\item{column_names}{A list of all of the column names}

\item{repeat_columns}{A list of all of the repeat
loops contained in the RHoMIS dataset}
}
\value{
A cleaned list of all of the column names
}
\description{
A single functions to shorten and
clean all RHoMIS column names.
}
\details{
Rpackage file: CleanColumnNames.R
}
\examples{

repeat_columns <- c(
  "crop_repeat",
  "livestock_repeat",
  "offfarm_repeat",
  "hh_rep"
)
seperator <- "/"
column_names <- c(
  "xxx/crop_repeat[1]/crop_name",
  "xxx/livestock_repeat[2]/livestock_name",
  "xxx/crop_repeat[3]/crop_name",
  "xx/crop_repeat/crop_name",
  "x/offfarm_repeat[4]/offfarm_name",
  "y/hh_rep[5]/person_name",
  "z/crop_repeat/crop_name"
)

# Will return:
# c("crop_name_1",
# "livestock_name_2",
# "crop_name_3",
# "crop_name",
# "offfarm_name_4",
# "person_name_5",
# "crop_name")
}
