% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LivestockCalculations.R
\name{clean_tlu_column_names}
\alias{clean_tlu_column_names}
\title{Clean TLU Column Names}
\usage{
clean_tlu_column_names(
  data,
  livestock_name_conversion_tibble,
  livestock_tlu_conversions
)
}
\arguments{
\item{data}{A rhomis dataset}

\item{livestock_name_conversion_tibble}{A tibble to convert
livestock names}

\item{livestock_tlu_conversions}{A conversions table for TLU}
}
\description{
TLU column names include options for
"other", this is not interperatable or
easy to use for analysis. This function addresses
this
}
\details{
Rpackage file: LivestockCalculations.R
}
