% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CropCalculations.R
\name{convert_crop_sold_units}
\alias{convert_crop_sold_units}
\title{Convert Crop Sold Units}
\usage{
convert_crop_sold_units(data, unit_conv_tibble = NULL)
}
\arguments{
\item{data}{RHoMIS data containing crop-looping information}

\item{unit_conv_tibble}{A tibble of unit conversions.}
}
\value{

}
\description{
Convert crop income units into a numeric conversion factor
}
