% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CropCalculations.R
\name{convert_crop_yield_units}
\alias{convert_crop_yield_units}
\title{Convert Crop Yield Units}
\usage{
convert_crop_yield_units(data, unit_conversions = crop_yield_units)
}
\arguments{
\item{data}{The data containing the units to convert}

\item{unit_conversions}{A dataframe or tibble of unit conversions}
}
\value{

}
\description{
RHoMIS crop yields come in a character format. This function
converts these to numeric and adds the new numeric units in an
appropriate place for the dataset
}
