% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CropCalculations.R
\name{convert_crop_yield_units}
\alias{convert_crop_yield_units}
\title{Convert Crop Yield Units}
\usage{
convert_crop_yield_units(
  data,
  units = crop_yield_units$unit,
  unit_conversions = crop_yield_units$conversion
)
}
\arguments{
\item{data}{The data containing the units to convert}

\item{units}{Units for crop yield}

\item{unit_conversions}{Conversions to convert these units
to kg}
}
\value{

}
\description{
RHoMIS crop yields come in a character format. This function
converts these to numeric and adds the new numeric units in an
appropriate place for the dataset
}
