% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhomisApiConnection.R
\name{create_project}
\alias{create_project}
\title{Create Project}
\usage{
create_project(api_url, auth_url, email, password, name)
}
\arguments{
\item{api_url}{The url of the rhomis data api}

\item{auth_url}{The url of the rhomis authentication api}

\item{email}{The users email}

\item{password}{The users password}

\item{name}{The name of the project to create}
}
\description{
Create a project in the RHoMIS database
}
