% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CropCalculations.R
\name{crop_calculations_all}
\alias{crop_calculations_all}
\title{Crop Calculations All}
\usage{
crop_calculations_all(
  data,
  crop_yield_units_conv_tibble = crop_yield_units,
  crop_income_units_conv_tibble = crop_price_units,
  gender_categories = pkg.env$gender_categories
)
}
\arguments{
\item{data}{RHoMIS crop loop data}

\item{crop_yield_units_conv_tibble}{Conversion tibble of crop yield units}

\item{crop_income_units_conv_tibble}{Conversion tibble of crop income units}

\item{gender_catgegories}{The categories you are interested in examining}
}
\value{

}
\description{
A single function for conducting all of the crop calculations
}
