% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CropCalculations.R
\name{crop_gender_calculations}
\alias{crop_gender_calculations}
\title{Crop Gender Calculations}
\usage{
crop_gender_calculations(
  data,
  gender_categories = c("female_youth", "female_adult", "male_youth", "male_adult")
)
}
\arguments{
\item{data}{RHoMIS data set conatining processed values for
crop consumed, crop sold, and crop income}

\item{gender_categories}{The categories you are interested in examining}
}
\value{

}
\description{
Adding the gendered information for crop consumed
crop sold, and crop income. These have to already have been
calculated in the dataset being used
}
