% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CropCalculations.R
\name{crop_harvest_calculations}
\alias{crop_harvest_calculations}
\title{Crop harvest calculations}
\usage{
crop_harvest_calculations(data, unit_conv_tibble = NULL)
}
\arguments{
\item{data}{RHoMIS data containing cropping information}

\item{unit_conv_tibble}{Units and conversions to convert crop yields to kg}
}
\value{

}
\description{
Calculating the amount of each crop harvested per year
}
