% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CropCalculations.R
\name{crop_income_calculations}
\alias{crop_income_calculations}
\title{Crop Income Calculations}
\usage{
crop_income_calculations(data)
}
\arguments{
\item{data}{A RHoMIS dataset, including information on crop harvested,
and crop sold}
}
\value{

}
\description{
A function for calculating crop incomes. Please note
for this calculation to work. The amount of a crop harvested, and the amount
of a crop sold also needs to have been calculated
}
