% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CropCalculations.R
\name{crop_income_calculations}
\alias{crop_income_calculations}
\title{Crop Income Calculations}
\usage{
crop_income_calculations(data, unit_conv_tibble = NULL)
}
\arguments{
\item{data}{A RHoMIS dataset, including information on crop harvested,
and crop sold}

\item{unit_conv_tibble}{Units for crop income calculations which
need to be converted}
}
\description{
A function for calculating crop incomes. Please note
for this calculation to work. The amount of a crop harvested, and the amount
of a crop sold also needs to have been calculated
}
\details{
Rpackage file: CropCalculations.R
}
