% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LivestockCalculations.R
\name{egg_income_calculations}
\alias{egg_income_calculations}
\title{Eggs Income Calculations}
\usage{
egg_income_calculations(
  data,
  units = eggs_price_time_units$unit,
  unit_conversions = eggs_price_time_units$conversion_factor
)
}
\arguments{
\item{data}{RHoMIS data with livestock loops included.}

\item{units}{List of units}

\item{unit_conversions}{List of conversion factors for the units provided above}
}
\value{

}
\description{
Function to calculate egg income from livestock loops
}
