% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LivestockCalculations.R
\name{eggs_amount_calculations}
\alias{eggs_amount_calculations}
\title{EggsAmount Calculations}
\usage{
eggs_amount_calculations(
  data,
  units = eggs_amount_units$unit,
  unit_conversions = eggs_amount_units$conversion_factor
)
}
\arguments{
\item{unit_conversions}{}
}
\value{

}
\description{
Calculate the amount of eggs harvested
}
