% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LivestockCalculations.R
\name{eggs_amount_calculations}
\alias{eggs_amount_calculations}
\title{EggsAmount Calculations}
\usage{
eggs_amount_calculations(data, unit_conv_tibble = NULL)
}
\arguments{
\item{data}{Data containing crop loop information}

\item{unit_conv_tibble}{A tibble of units and conversion factor, and project IDs}
}
\description{
Calculate the amount of eggs harvested
  
Rpackage file: LivestockCalculations.R
}
