% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-extract-units.R
\name{extract_units_and_conversions_csv}
\alias{extract_units_and_conversions_csv}
\title{Extract Units and Conversions CSV}
\usage{
extract_units_and_conversions_csv(file_path)
}
\arguments{
\item{file_path}{Filepath to a RHoMIS Survey CSV}
}
\description{
Rpackage file: 01-extract-units.R
}
