% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redirectModules.R
\name{extract_values_local}
\alias{extract_values_local}
\title{Extract values}
\usage{
extract_values_local(
  base_folder = "./",
  file_path = "./raw-data/raw-data.csv",
  country_column = "country",
  unique_id_col = "_uuid",
  hh_id_col = NULL,
  id_type = "string",
  proj_id = NULL,
  form_id = NULL,
  overwrite = FALSE
)
}
\arguments{
\item{file_path}{The path to the raw-data rhomis file}

\item{country_column}{The column name for the variable "country".}

\item{unique_id_col}{The column name of the odk uuid, usually "_uuid"}

\item{hh_id_col}{The column name containing household IDs.}

\item{id_type}{Indicator of whether you are providing a single ID}

\item{proj_id}{Either a single string to be used as the project ID for all households, or the name of the column containing the project IDs (depending on id_type)}

\item{form_id}{Either a single string to be used as the form ID for all households, or the name of the column containing the form IDs (depending on id_type)}

\item{overwrite}{True if you would like to overwrite previous ID column, false if would not like to overwrite existing IDs}
}
\value{

}
\description{
Extract values
}
