% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redirectModules.R
\name{extract_values_local}
\alias{extract_values_local}
\title{Extract values
   
Rpackage file: redirectModules.R}
\usage{
extract_values_local(
  base_folder = "./",
  file_path = "./raw-data/raw-data.csv",
  country_column = pkg.env$identification_column_list$country,
  unique_id_col = pkg.env$identification_column_list$uuid_local,
  hh_id_col = NULL,
  id_type = c("string", "column"),
  proj_id,
  form_id,
  overwrite = FALSE,
  repeat_column_names = pkg.env$repeat_columns
)
}
\arguments{
\item{base_folder}{The folder where all outputs will be written to}

\item{file_path}{The path to the raw-data rhomis file}

\item{country_column}{The column name for the variable "country".}

\item{unique_id_col}{The column name of the odk uuid, usually "_uuid"}

\item{hh_id_col}{The column name containing household IDs.}

\item{id_type}{Indicator of whether you are providing a single ID}

\item{proj_id}{Either a single string to be used as the project ID for all households, or the name of the column containing the project IDs (depending on id_type)}

\item{form_id}{Either a single string to be used as the form ID for all households, or the name of the column containing the form IDs (depending on id_type)}

\item{overwrite}{True if you would like to overwrite previous ID column, false if would not like to overwrite existing IDs}

\item{repeat_column_names}{The prefixes for repeat columns (e.g. 'section/crop_repeat[1]/crop_name', the prefix would be crop_repeat)}
}
\description{
Extract values
   
Rpackage file: redirectModules.R
}
