% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FoodSecurity.R
\name{fies_score}
\alias{fies_score}
\title{fies score}
\usage{
fies_score(data)
}
\arguments{
\item{data}{The data set containing all of the RHoMIS data (or just the fies information)}
}
\description{
Rpackage file: FoodSecurity.R
}
