% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NTFP.R
\name{fp_proportions_all}
\alias{fp_proportions_all}
\title{FP Proportions All}
\usage{
fp_proportions_all(tree_aid_df, use, use_column, prop_column, new_column_name)
}
\arguments{
\item{new_column_name}{}
}
\description{
Calculate numeric proportions NTFPs sold and consumed
}
