% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TotalIncomes.R
\name{gendered_off_farm_income_split}
\alias{gendered_off_farm_income_split}
\title{Gendered Off-Farm Income Split}
\usage{
gendered_off_farm_income_split(data, gender_categories)
}
\arguments{
\item{data}{RHoMIS data including off-farm income loops}

\item{gender_categories}{Categories of gender we are interested in}
}
\description{
Adding gender control columns for each of the specific
sources of off-farm income
}
\details{
Rpackage file: TotalIncomes.R
}
