% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenerateMockResponses.R
\name{generate_mock_response}
\alias{generate_mock_response}
\title{Generate Mock Response}
\usage{
generate_mock_response(survey, choices, metadata, survey_path = NULL)
}
\arguments{
\item{survey}{The "survey" tab of the survey xls file}

\item{choices}{The "choices" tab of the survey xls file}

\item{metadata}{The "settings" tab of the survey xls file}

\item{survey_path}{The path to the xls survey file}
}
\description{
This is a function to generate a mock response based on
the survey file alone
}
\details{
Rpackage file: GenerateMockResponses.R
}
