% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-odk-central.R
\name{get_submission_data}
\alias{get_submission_data}
\title{Get Submission Data}
\usage{
get_submission_data(
  central_url,
  central_email,
  central_password,
  projectID,
  formID,
  form_name,
  isDraft,
  file_destination = NULL
)
}
\arguments{
\item{central_url}{The url of the ODK central server}

\item{central_email}{The email of your ODK central account}

\item{central_password}{The password to your ODK central account}

\item{projectID}{The ID of the project you are looking at.
To get a list of project, see the
"get_projects" function}

\item{formID}{The ID of the form containing the
submissions you are looking at.}

\item{form_name}{The name of the form you are downloading
To get the list of forms see the "get_forms" function.}

\item{isDraft}{Whether or not the form is a draft or whether it is finalized}

\item{file_destination}{The location to store temporary files}
}
\description{
A function for retrieving all of the ODK data as a zip file
}
\details{
Rpackage file: utils-odk-central.R
}
