% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConnectToCentral.R
\name{get_submission_data}
\alias{get_submission_data}
\title{Get Submission Data}
\usage{
get_submission_data(
  central_url,
  central_email,
  central_password,
  projectID,
  formID,
  draft,
  file_destination = NULL,
  central_test_case = F
)
}
\arguments{
\item{central_url}{The url of the ODK central server}

\item{central_email}{The email of your ODK central account}

\item{central_password}{The password to your ODK central account}

\item{projectID}{The ID of the project you are looking at.
To get a list of project, see the
"get_projects" function}

\item{formID}{The ID of the form containing the
submissions you are looking at.
To get the list of forms see the "get_forms" function.}

\item{draft}{Whether or not the form is a draft or whether it is finalized}

\item{file_destination}{The location to store temporary files}

\item{central_test_case}{Whether or not you are running a test
example to show how data downloads from ODK central work}
}
\value{

}
\description{
A function for retrieving all of the ODK data as a zip file
}
