% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LivestockCalculations.R
\name{honey_amount_calculation}
\alias{honey_amount_calculation}
\title{Honey Amount Calculation}
\usage{
honey_amount_calculation(
  data,
  units = honey_amount_units$units,
  unit_conversions = honey_amount_units$conversion_factors
)
}
\arguments{
\item{data}{The data containing livestock loops}

\item{units}{A vector of units for which you have conversion factors}

\item{unit_conversions}{The conversion factors for those units}
}
\value{

}
\description{
Calculating the amount of honey produced from RHoMIS data
}
