% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LivestockCalculations.R
\name{honey_income_calculations}
\alias{honey_income_calculations}
\title{Calculate the prices for honey
  
Rpackage file: LivestockCalculations.R}
\usage{
honey_income_calculations(data)
}
\arguments{
\item{data}{RHoMIS dataset.}
}
\description{
Calculate the prices for honey
  
Rpackage file: LivestockCalculations.R
}
