% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LivestockCalculations.R
\name{honey_proportions_all}
\alias{honey_proportions_all}
\title{Honey Proportions All}
\usage{
honey_proportions_all(data)
}
\arguments{
\item{data}{The data containing livestock loops}
}
\description{
Can correctly calculate the numeric proportions of honey
and its uses
}
\details{
Rpackage file: LivestockCalculations.R
}
