% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HouseholdSize.R
\name{household_roster_to_categories}
\alias{household_roster_to_categories}
\title{Household roster to categories}
\usage{
household_roster_to_categories(data)
}
\arguments{
\item{data}{The dataset containing the household roster information}
}
\value{
A data-frame with the correct household
}
\description{
Rpackage file: HouseholdSize.R
}
