% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LandSize.R
\name{land_size_calculation}
\alias{land_size_calculation}
\title{Land Size Calculation}
\usage{
land_size_calculation(data, unit_conv_tibble = NULL)
}
\arguments{
\item{data}{RHoMIS data}

\item{unit_conv_tibble}{A tibble of land area units and}
}
\value{

}
\description{
Calculate land sizes in hectares using
RHoMIS data
}
