% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LivestockCalculations.R
\name{livestock_calculations_all}
\alias{livestock_calculations_all}
\title{Livestock calculations all}
\usage{
livestock_calculations_all(
  data,
  livestock_weights_conv_tibble = NULL,
  eggs_amount_unit_conv_tibble = NULL,
  eggs_price_time_units_conv_tibble = NULL,
  honey_amount_unit_conv_tibble = NULL,
  milk_amount_unit_conv_tibble = NULL,
  milk_price_time_unit_conv_tibble = NULL,
  gender_categories = pkg.env$gender_categories
)
}
\arguments{
\item{data}{RHoMIS data, containing livestock loops}

\item{livestock_weights_conv_tibble}{Conversion tibble of livestock weights}

\item{eggs_amount_unit_conv_tibble}{Conversion tibble of eggs amount}

\item{eggs_price_time_units_conv_tibble}{Conversion tibble of egg price time units}

\item{honey_amount_unit_conv_tibble}{Conversion tibble of honey amount}

\item{milk_amount_unit_conv_tibble}{Conversion tibble of milk amounts}

\item{milk_price_time_unit_conv_tibble}{Conversion tibble of milk prices}

\item{gender_categories}{The gender control catgeories included in the survey}
}
\description{
Carrying out all calculations on RHoMIS
livestock loops.
}
\details{
Rpackage file: LivestockCalculations.R
}
