% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{livestock_tlu}
\alias{livestock_tlu}
\title{Livestock TLU Conversions}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 15 rows and 2 columns.
}
\source{
NA
}
\usage{
livestock_tlu
}
\description{
A table to convert different
types of livestock into
tropical livestock units
}
\keyword{datasets}
