% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UnitsAndConversions.R
\name{load_local_units}
\alias{load_local_units}
\title{Load local units}
\usage{
load_local_units(base_folder, id_rhomis_dataset)
}
\arguments{
\item{base_folder}{The path to the folder containing the units to load}

\item{id_rhomis_dataset}{A vector including the ID of the RHoMIS datasets being processed}
}
\value{

}
\description{
Load units for a particular project from csv and
load them into the global environment
}
