% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UnitsAndConversions.R
\name{load_local_units}
\alias{load_local_units}
\title{Load local units}
\usage{
load_local_units()
}
\value{

}
\description{
Load units for a particular project from csv and
load them into the global environment
}
