% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-odk-central.R
\name{load_rhomis_central}
\alias{load_rhomis_central}
\title{Load RHoMIS Central}
\usage{
load_rhomis_central(
  central_url = NULL,
  central_email = NULL,
  central_password = NULL,
  project_name = NULL,
  form_name = NULL,
  database = NULL,
  isDraft = NULL,
  central_test_case = FALSE
)
}
\arguments{
\item{central_url}{The url of the ODK-central server you are using.
ONLY RELEVANT IF "dataSource" WAS "central".}

\item{central_email}{The email of the ODK-central account you are using.
ONLY RELEVANT IF "dataSource" WAS "central"}

\item{central_password}{The password of the ODK-central account you are using.
ONLY RELEVANT IF "dataSource" WAS "central".}

\item{project_name}{The name of the ODK-central project you are processing.
ONLY RELEVANT IF "dataSource" WAS "central".}

\item{form_name}{The name of the ODK-central form you are processing.
ONLY RELEVANT IF "dataSource" WAS "central".}

\item{database}{The name of the database you would like to save results to}

\item{isDraft}{Whether or not the ODK form you are working with is a draft
or a final version. Only relevant if you are processing a project from ODK central}

\item{central_test_case}{This flag is used for running a test-sample dataset from ODK the inst/sample_central_project/ folder}
}
\description{
Load Raw RHoMIS data from ODK central and
convert the column names into a shortened, standardised
version.
}
\details{
Rpackage file: utils-odk-central.R
}
