% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunAll.R
\name{load_units_csvs}
\alias{load_units_csvs}
\title{Load Units CSVs}
\usage{
load_units_csvs(folder, ids_rhomis_dataset)
}
\arguments{
\item{folder}{The folder containing all of the unit conversions.}

\item{id_rhomis_dataset}{A vector of the RHoMIS IDs}
}
\value{

}
\description{
Load Units CSVs
}
