% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunAll.R
\name{make_id_columns}
\alias{make_id_columns}
\title{Make ID Columns}
\usage{
make_id_columns(
  data,
  country_column = "country",
  unique_id_col = "_uuid",
  hh_id_col = NULL,
  id_type = c("string", "column"),
  proj_id,
  form_id
)
}
\arguments{
\item{data}{The rhomis data set as a tibble}

\item{country_column}{The name of the column containing the country string, as collected in the survey}

\item{unique_id_col}{The name of the column containing unique id record. This is produced by the server accepting ODK records}

\item{hh_id_col}{The household ID column}

\item{id_type}{The type of ID you would like to enter for projects and forms. If you select "string", then fill in the proj_id and form_id arguments, with the project id and form id you would like to use. If selecting "column", enter the name of the column (proj_id) containing the project ID you would like to use, and the name of the column (form_id) containing the form ids you would like to use.}

\item{proj_id}{Either a single string to be used as the project ID for all households, or the name of the column containing the project IDs (depending on id_type)}

\item{form_id}{Either a single string to be used as the form ID for all households, or the name of the column containing the form IDs (depending on id_type)}
}
\description{
Make ID columns for form, project, and household
 
Rpackage file: RunAll.R
}
