% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LivestockCalculations.R
\name{meat_amount_calculation}
\alias{meat_amount_calculation}
\title{Meat amount calculations}
\usage{
meat_amount_calculation(data, unit_conv_tibble = NULL)
}
\arguments{
\item{unit_conv_tibble}{A conversion table for animal names and their weights}

\item{animal_weights_names}{The vector which contains the names of the animals}
}
\value{

}
\description{
A function to calculate the amount of meat
produced based on the number of animals killed,
and some conversion factors which can calculate
how much meat can be collected from an animal which
has been killed.
}
