% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LivestockCalculations.R
\name{meat_amount_calculation}
\alias{meat_amount_calculation}
\title{Meat amount calculations}
\usage{
meat_amount_calculation(data, unit_conv_tibble = NULL)
}
\arguments{
\item{data}{The dataset for which meat amounts need to be calculated}

\item{unit_conv_tibble}{A conversion table for animal names and their weights}
}
\description{
A function to calculate the amount of meat
produced based on the number of animals killed,
and some conversion factors which can calculate
how much meat can be collected from an animal which
has been killed.
  
Rpackage file: LivestockCalculations.R
}
