% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LivestockCalculations.R
\name{milk_amount_calculations}
\alias{milk_amount_calculations}
\title{Milk amount calculations}
\usage{
milk_amount_calculations(data, unit_conv_tibble = NULL)
}
\arguments{
\item{data}{The dataset containing livestock loops for RHoMIS}

\item{unit_conv_tibble}{The units for milk amounts in a conversion tibble}
}
\description{
Calculating the amount of milk collected each year
  
Rpackage file: LivestockCalculations.R
}
