% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LivestockCalculations.R
\name{milk_income_calculations}
\alias{milk_income_calculations}
\title{Milk Income Calculations}
\usage{
milk_income_calculations(
  data,
  units = milk_price_time_units$unit,
  conversion_factors = milk_price_time_units$conversion_factor
)
}
\arguments{
\item{data}{Dataset containing all milk income information}

\item{units}{The names of the units you hope to convert}

\item{conversion_factors}{The conversion factors for these units}
}
\value{

}
\description{
Milk Income Calculations
}
