% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LivestockCalculations.R
\name{milk_price_time_units_conversion}
\alias{milk_price_time_units_conversion}
\title{Milk Price Time Units Conversion}
\usage{
milk_price_time_units_conversion(
  id_rhomis_dataset,
  units_column,
  sold_amount_column,
  unit_conv_tibble = NULL
)
}
\arguments{
\item{id_rhomis_dataset}{An id vector containing information on which rhomis}

\item{units_column}{The columns containing the original units}

\item{sold_amount_column}{The amounts of milk which was sold}

\item{unit_conv_tibble}{A table with all of the information to convert price time units}
}
\description{
RHoMIS milk price units can come in both times and per
litre units. This function goes
  
Rpackage file: LivestockCalculations.R
}
