% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NTFP.R
\name{ntfp_preprocessing}
\alias{ntfp_preprocessing}
\title{NTFP Preprocessing}
\usage{
ntfp_preprocessing(tree_aid_df)
}
\arguments{
\item{tree_aid_df}{A dataset containing (or not) questions from NTFP modules}
}
\description{
For forest product surveys, honey
does not fit into the conventional NTFP loops.
This means we cannot make use of existing NTFP calculations.
}
\details{
This script adds an "extra loop" to a NTFP dataset, one
that allows for the calculation of honey based-indicators.
}
