% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NTFP.R
\name{ntfp_sold_and_consumed_calculation}
\alias{ntfp_sold_and_consumed_calculation}
\title{NTFP sold and consumed Calculation}
\usage{
ntfp_sold_and_consumed_calculation(
  rhomis_data,
  fp_harvest_kg,
  fp_props_sold_numeric,
  fp_amount_sold_kg,
  fp_prop_consumed_numeric,
  fp_prop_consumed_kg
)
}
\arguments{
\item{fp_prop_consumed_kg}{}
}
\description{
Calculate the amounts of NTFPs sold and
consumed in KG
}
