% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NTFP.R
\name{ntfp_totals}
\alias{ntfp_totals}
\title{NTFP Totals}
\usage{
ntfp_totals(tree_aid_df, indicator_df, fp_products = fp_products)
}
\arguments{
\item{tree_aid_df}{Tree aid data frame}

\item{indicator_df}{Data frame containing table of indicators}

\item{fp_products}{List of all of the forest products}
}
\description{
Calculate total ntfp income, value consumed,
calories consumed, processed income, value
of processed ntfp consumed, processed
calories consumed
}
