% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunAll.R
\name{processData}
\alias{processData}
\title{Process RHoMIS data}
\usage{
processData(
  extractUnitsOnly = T,
  calculateInitialIndicatorsOnly = F,
  calculateFinalIndicatorsOnly = F,
  dataSource = c("csv", "central"),
  outputType = c("csv", "mongodb"),
  base_path = "./",
  dataFilePath = NULL,
  id_type = c("string", "column"),
  proj_id,
  form_id,
  uuid_local = pkg.env$identification_column_list$uuid_local,
  central_url = NULL,
  central_email = NULL,
  central_password = NULL,
  project_name = NULL,
  form_name = NULL,
  form_version = NULL,
  database = NULL,
  isDraft = NULL,
  central_test_case = FALSE,
  repeat_columns = pkg.env$repeat_columns,
  gender_categories = pkg.env$gender_categories
)
}
\arguments{
\item{extractUnitsOnly}{Whether or not to only extract units (TRUE/FALSE)}

\item{calculateInitialIndicatorsOnly}{Whether or not to only calculate
initial indicators (TRUE/FALSE)}

\item{calculateFinalIndicatorsOnly}{Whether or not to only calculate
final indicators}

\item{dataSource}{The type of RHoMIS data being fed into the
calculations, whether a local csv file or data from ODK central.
Options "csv" or "central".}

\item{outputType}{The type of output to produce (options are "csv"
or "mongodb")}

\item{base_path}{The path where all of the data processing should take place}

\item{dataFilePath}{The file to the data (csv format).
ONLY RELEVANT IF "dataSource" WAS "local".}

\item{id_type}{The type of ID you would like to use ("string" or "column")}

\item{proj_id}{An ID for your project}

\item{form_id}{An ID for your form}

\item{uuid_local}{The column in a local dataset containing uuids (usually _uuid)}

\item{central_url}{The url of the ODK-central server you are using.
ONLY RELEVANT IF "dataSource" WAS "central".}

\item{central_email}{The email of the ODK-central account you are using.
ONLY RELEVANT IF "dataSource" WAS "central"}

\item{central_password}{The password of the ODK-central account you are using.
ONLY RELEVANT IF "dataSource" WAS "central".}

\item{project_name}{The name of the ODK-central project you are processing.
ONLY RELEVANT IF "dataSource" WAS "central".}

\item{form_name}{The name of the ODK-central form you are processing.
ONLY RELEVANT IF "dataSource" WAS "central".}

\item{form_version}{The version of the ODK-central form you are processing.
ONLY RELEVANT IF "dataSource" WAS "central".}

\item{database}{The name of the database you would like to save results to}

\item{isDraft}{Whether or not the ODK for you war working with is a draft
or a final version. Only relevant if you are processing a project from ODK central}

\item{central_test_case}{This flag is used for running a test-sample dataset from ODK the inst/sample_central_project/ folder}

\item{repeat_columns}{The columns which are looped in the datasets being processed}

\item{gender_categories}{The gender categories present in the data which is to be processed}
}
\value{

}
\description{
A main function that can be used to process rhomis
data. Whether the dataset comes from a local csv
or from ODK central.
}
\details{
RHoMIS datasets go through 4 stages of processing:

1. Initial Cleaning and Extracting Units
2. Calculation of initial indicators
3. Calculation of final indicators, including food availability,
gender, and value of products consumed
}
