% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunAll.R
\name{processData}
\alias{processData}
\title{Process RHoMIS data}
\usage{
processData(
  dataSource = "csv",
  outputType = "csv",
  coreOnly = T,
  surveyFile = NULL,
  moduleSaving = F,
  extractUnits = T,
  calculatePrices = F,
  processData = F,
  dataFilePath = NULL,
  central_url = NULL,
  central_email = NULL,
  central_password = NULL,
  project_name = NULL,
  form_name = NULL,
  form_version = NULL,
  database = NULL
)
}
\arguments{
\item{dataSource}{The type of RHoMIS data being fed into the
calculations, whether a local csv file or data from ODK central.
Options "local" or "central".}

\item{outputType}{The type of output to produce (options are "csv" or "mongodb")}

\item{coreOnly}{Indicating whether to only analyze the RHoMIS core information (TRUE/FALSE)
If TRUE, extra variables will be saved but they will not be processed into indicators.
If FALSE, you must either provide the survey xls file so that the modules used can be identified,
or ensure that the project you are accessing can be found on ODK central.}

\item{surveyFile}{The path to the surveyxls file. Only necessary if "coreOnly"=FALSE and dataSource="local".}

\item{moduleSaving}{Whether or not to use the form to identify the modules which
were included in the survey (TRUE/FALSE)}

\item{extractUnits}{Whether or not to only extract units (TRUE/FALSE)}

\item{calculatePrices}{Whether or not to only calculate the new prices (TRUE/FALSE)}

\item{processData}{Whether to process the whole dataset (TRUE/FALSE)}

\item{dataFilePath}{The file to the data (csv format).
ONLY RELEVANT IF "dataSource" WAS "local".}

\item{central_url}{The url of the ODK-central server you are using.
ONLY RELEVANT IF "dataSource" WAS "central".}

\item{central_email}{The email of the ODK-central account you are using.
ONLY RELEVANT IF "dataSource" WAS "central"}

\item{central_password}{The password of the ODK-central account you are using.
ONLY RELEVANT IF "dataSource" WAS "central".}

\item{project_name}{The name of the ODK-central project you are processing.
ONLY RELEVANT IF "dataSource" WAS "central".}

\item{form_name}{The name of the ODK-central form you are processing.
ONLY RELEVANT IF "dataSource" WAS "central".}

\item{form_version}{The version of the ODK-central form you are processing.
ONLY RELEVANT IF "dataSource" WAS "central".}

\item{database}{The name of the database you would like to save results to}
}
\value{

}
\description{
A main function that can be used to process rhomis
data. Whether the dataset comes from a local csv
or from ODK central
}
