% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AuthenticateUser.R
\name{register_user}
\alias{register_user}
\title{These functions are helpful for managing
users onf the RHoMIS 2.0 authentication server.
Found here: https://github.com/l-gorman/rhomis-authenticator
Register User}
\usage{
register_user(url, username, email, password)
}
\arguments{
\item{url}{The url of the server we are using for authentication}

\item{username}{The username of the user to be registered}

\item{email}{The email of the user to be registered}

\item{password}{The password of the user to be registered}
}
\value{
Will return the user ID from the database
}
\description{
Register a user on the RHoMIS 2.0
authentication server
}
\details{
Rpackage file: AuthenticateUser.R
}
