% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CleanRawData.R
\name{switch_units}
\alias{switch_units}
\title{Switch Units}
\usage{
switch_units(data_to_convert, unit_tibble, id_vector)
}
\arguments{
\item{data_to_convert}{This list or data frame which needs to be converted}

\item{unit_tibble}{A tibble containing}

\item{id_vector}{Column containing the RHoMIS id, used to subset the units by project data frame}
}
\value{
Returns the data (in its original format) with the units in their numerical format
}
\description{
In RHoMIS, sometimes lists or data frames contain units which need to be converted
from a character format (e.g. "kg","sacks_100kg") into a numeric conversion (e.g. 1, 100).
This function does just that.
}
\examples{
# Converting a list

}
