% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ValueCalculations.R
\name{value_calculations}
\alias{value_calculations}
\title{Value Calculations}
\usage{
value_calculations(processed_data, indicator_data, prices)
}
\arguments{
\item{mean_prices}{}
}
\value{

}
\description{
Value Calculations
}
