% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ValueCalculations.R
\name{value_calculations}
\alias{value_calculations}
\title{Value Calculations}
\usage{
value_calculations(processed_data, indicator_data, prices, gender_categories)
}
\arguments{
\item{processed_data}{A processed rhomis dataset}

\item{indicator_data}{An indicator tibble}

\item{prices}{A list of tibbles including prices}

\item{gender_categories}{The gender categories to examing}
}
\value{

}
\description{
Calculations of the value of products consumed
}
