% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redirectModules.R
\name{value_gender_fa_calculations}
\alias{value_gender_fa_calculations}
\title{Value, Gender, and Food Availability Calculations}
\usage{
value_gender_fa_calculations(
  processed_data,
  indicator_data,
  calorie_conversions,
  prices,
  gender_categories
)
}
\arguments{
\item{processed_data}{RHoMIS Processed Dataset}

\item{indicator_data}{RHoMIS Indicator Data set}

\item{calorie_conversions}{Calorie conversions list}

\item{prices}{A list of tibbles on mean prices}

\item{gender_categories}{Gender categories to convert}
}
\value{

}
\description{
Value, Gender, and Food Availability Calculations
}
